LDPC Codes

Constructors

CodingTheory.LDPCCode — Type

LDPCCode(H::fq_nmod_mat)

Return the LDPC code defined by the parity-check matrix H.

Notes

LDPC codes are typically required to have a matrix density of less than 1%.

LDPCCode(C::AbstractLinearCode)

Return the LDPC code given by C.

Notes

LDPC codes are typically required to have a matrix density of less than 1%.

Attributes

CodingTheory.variabledegreedistribution — Function

variabledegreedistribution(C::AbstractLDPCCode)

Return the variable node degree distribution of C.

CodingTheory.checkdegreedistribution — Function

checkdegreedistribution(C::AbstractLDPCCode)

Return the check node degree distribution of C.

CodingTheory.degreedistributions — Function

degreedistributions(C::AbstractLDPCCode)

Return the variable and check node degree distributions of C.

CodingTheory.variabledegreepolynomial — Function

variabledegreepolynomial(C::AbstractLDPCCode)

Return the variable degree polynomial of C.

CodingTheory.checkdegreepolynomial — Function

checkdegreepolynomial(C::AbstractLDPCCode)

Return the check degree polynomial of C.

CodingTheory.columnbound — Function

columnbound(C::AbstractLDPCCode)

Return the column bound c of the (c, r)-LDPC code C.

CodingTheory.rowbound — Function

rowbound(C::AbstractLDPCCode)

Return the row bound r of the (c, r)-LDPC code C.

CodingTheory.columnrowbounds — Function

columnrowbounds(C::AbstractLDPCCode)

Return the column and row bounds c, r of the (c, r)-LDPC code C.

CodingTheory.limited — Function

limited(C::AbstractLDPCCode)

Return the maximum of the row and column bounds for C.

Hecke.density — Function

density(C::AbstractLDPCCode)

Return the density of the parity-check matrix of C.

Hecke.isregular — Function

isregular(G::SimpleGraph{Int})

Return true if G is regular.

isregular(C::AbstractLDPCCode)

Return true if the C is a regular LDPC code.

Notes

An LDPC is regular if all the column degrees and equal and all the row degrees are equal.

Methods

CodingTheory.degreedistributionssplot — Function

degreedistributionssplot(C::AbstractLDPCCode)

Return a bar plot of the column and row degree distributions of C.

Tanner

CodingTheory.Tannergraph — Function

Tannergraph(H::Union{fq_nmod_mat, Matrix{Int}})

Return the SimpleGraph object repesenting the Tanner graph of the parity-check matrix H along with the indices of the left and right vertices representing the bits and parity checks, respectively.

Tannergraph(C::AbstractLinearCode)

Return the SimpleGraph object repesenting the Tanner graph of C along with the indices of the left and right vertices representing the bits and parity checks, respectively.

Tannergraph(C::AbstractLDPCCode)

Return the Tanner graph of C as a Figure object.

CodingTheory.Tannergraphplot — Function

Tannergraphplot(H::Union{fq_nmod_mat, Matrix{Int}})

Return the Tanner graph of the matrix H as a Figure object.

CodingTheory.Tannercode — Function

Tannercode(EVI::SparseMatrixCSC{Int, Int}, C::AbstractLinearCode)

Return the Tanner code obtained by applying the local code C to the edges of the graph with edge-vertex incidence matrix EVI.

Tannercode(G::SimpleGraph{Int}, C::AbstractLinearCode)

Return the Tanner code obtained by applying the local code C to the edges of G.

Tannercode(G::SimpleGraph{Int}, left::Vector{Int}, right::Vector{Int}, C::AbstractLinearCode)

Return the Tanner code obtained by applying the local code C to the vertices right in the bipartition of G and treating the vertices of left as bits.

Tannercode(G::SimpleGraph{Int}, left::Vector{Int}, right1::Vector{Int}, right2::Vector{Int}, C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the Tanner code obtained by applying the local code C1 to the vertices right1 and the local code C2 to the vertices right2 in the bipartition of G and treating the vertices of left as bits.

LDPC Construction Algorithms