Cyclotomic Cosets

CodingTheory.allcyclotomiccosets — Function

allcyclotomiccosets(q::Int, n::Int, tosort::Bool=true, verbose::Bool=false)

Return all q-cyclotomic cosets modulo n.

Notes

If the optional parameter tosort is set to false, the result will not be

sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.complementqcosets — Method

complementqcosets(q::Int, n::Int, qcosets::Vector{Vector{Int64}})

Return the complement of the q-cyclotomic cosets modulo n of qcosets.

CodingTheory.cyclotomiccoset — Function

cyclotomiccoset(x::Int, q::Int, n::Int, tosort::Bool=true, verbose::Bool=false)

Return the q-cyclotomic coset of x modulo n.

Notes

If the optional parameter tosort is set to false, the result will not be

sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.dualqcosets — Method

dualqcosets(q::Int, n::Int, qcosets::Vector{Vector{Int64}})

Return the dual of the q-cyclotomic cosets modulo n of qcosets.

CodingTheory.ord — Method

ord(n::Int, q::Int)

Return the order of n mod q.

CodingTheory.qcosetpairings — Method

qcosetpairings(arr::Vector{Vector{Int64}}, n::Int)

Return the q-cyclotomic cosets modulo n collected into complementary pairs.

Cyclic Codes

CodingTheory.BCHCode — Type

BCHCode(q::Int, n::Int, δ::Int, b::Int=0)

Return the BCHCode of length n over GF(q) with design distance δ and offset b.

Notes

This function will auto determine if the constructed code is Reed-Solomon

and call the appropriate constructor.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> B = BCHCode(q, n, δ, b)
[15, 5, ≥7; 1]_2 BCH code over splitting field GF(16).
2-Cyclotomic cosets:
        C_1 ∪ C_3 ∪ C_5
Generator polynomial:
        x^10 + x^8 + x^5 + x^4 + x^2 + x + 1
Generator matrix: 5 × 15
        1 1 1 0 1 1 0 0 1 0 1 0 0 0 0
        0 1 1 1 0 1 1 0 0 1 0 1 0 0 0

        0 0 1 1 1 0 1 1 0 0 1 0 1 0 0
        0 0 0 1 1 1 0 1 1 0 0 1 0 1 0
        0 0 0 0 1 1 1 0 1 1 0 0 1 0 1

CodingTheory.CyclicCode — Method

CyclicCode(q::Int, n::Int, cosets::Vector{Vector{Int}})

Return the CyclicCode of length n over GF(q) with q-cyclotomic cosets cosets.

Notes

This function will auto determine if the constructed code is BCH or Reed-Solomon

and call the appropriate constructor.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> cosets = definingset([i for i = b:(b + δ - 2)], q, n, false);
julia> C = CyclicCode(q, n, cosets)

CodingTheory.CyclicCode — Method

CyclicCode(n::Int, g::fq_nmod_poly)

Return the length n cyclic code generated by the polynomial g.

CodingTheory.QuasiCyclicCode — Type

QuasiCyclicCode(v::fq_nmod_mat, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l generated by right-bit shifts of size l of the generator vector v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

CodingTheory.QuasiCyclicCode — Type

QuasiCyclicCode(v::Vector{AbstractCyclicCode}, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l whose circulants are determined by the cyclic codes in v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

CodingTheory.QuasiCyclicCode — Method

QuasiCyclicCode(A::MatElem{T}, parity::Bool=false) where T <: ResElem

Return the quasi-cycle code specified by the matrix A of polynomial circulant generators. If the optional paramater parity is set to true, the input is used to construct the parity-check matrix.

CodingTheory.QuasiCyclicCode — Method

QuasiCyclicCode(v::Vector{fq_nmod_mat}, l::Int, circgens::Bool, parity::Bool=false)

Return the quasi-cyclic code of index l generated by right-bit shifts of size l of the generator vectors v. If circgens is true, the vectors are taken to be (column) generators for the circulant matrices instead of generator vectors for the code. If the optional paramater parity is set to true, the input is used to construct the parity-check matrix.

Notes

If circgens is false, then the length of the code is ncols(v[1]) and must be divisible by l.

If circgens is true, then the length of the code is ncols(v[1]) * l. Circulant matrices are stacked in rows of length l, so l must divide length(v).

CodingTheory.QuasiCyclicCode — Method

QuasiCyclicCode(v::Vector{fq_nmod_poly}, n::Int, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l whose circulants are defined by the generator polynomials v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

CodingTheory.ReedSolomonCode — Type

ReedSolomonCode(q::Int, δ::Int, b::Int=0)

Return the ReedSolomonCode over GF(q) with distance d and offset b.

Examples

julia> ReedSolomonCode(8, 3, 0)
[7, 5, ≥3; 0]_8 Reed Solomon code.
8-Cyclotomic cosets:
        C_0 ∪ C_1
Generator polynomial:
        x^2 + (α + 1)*x + α
Generator matrix: 5 × 7
        α α + 1 1 0 0 0 0
        0 α α + 1 1 0 0 0
        0 0 α α + 1 1 0 0
        0 0 0 α α + 1 1 0
        0 0 0 0 α α + 1 1

julia> ReedSolomonCode(13, 5, 1)
[12, 8, ≥5; 1]_13 Reed Solomon code.
13-Cyclotomic cosets:
        C_1 ∪ C_2 ∪ C_3 ∪ C_4
Generator polynomial:
        x^4 + 9*x^3 + 7*x^2 + 2*x + 10
Generator matrix: 8 × 12
        10 2 7 9 1 0 0 0 0 0 0 0
        0 10 2 7 9 1 0 0 0 0 0 0
        0 0 10 2 7 9 1 0 0 0 0 0
        0 0 0 10 2 7 9 1 0 0 0 0
        0 0 0 0 10 2 7 9 1 0 0 0
        0 0 0 0 0 10 2 7 9 1 0 0
        0 0 0 0 0 0 10 2 7 9 1 0
        0 0 0 0 0 0 0 10 2 7 9 1

Base.:+ — Method

+(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the addition code of C1 and C2.

Base.:== — Method

==(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return whether or not C1 and C2 have the same fields, lengths, and defining sets.

Base.:∩ — Method

∩(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the intersection code of C1 and C2.

Base.:⊆ — Method

⊆(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
⊂(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
issubcode(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return whether or not C1 is a subcode of C2.

Base.zeros — Method

zeros(C::AbstractCyclicCode)

Return the zeros of C.

CodingTheory.BCHbound — Method

BCHbound(C::AbstractCyclicCode)

Return the BCH bound for C.

Notes

This is a lower bound on the minimum distance of C.

CodingTheory.circulants — Method

circulants(C::AbstractQuasiCyclicCode)

Return the circulant matrices of the quasi-cyclic code.

CodingTheory.definingset — Function

definingset(nums::Vector{Int}, q::Int, n::Int, flat::Bool=true)

Returns the set of q-cyclotomic cosets of the numbers in nums modulo n.

Notes

If flat is set to true, the result will be a single flattened and sorted array.

CodingTheory.definingset — Method

definingset(C::AbstractCyclicCode)

Return the defining set of the cyclic code.

CodingTheory.designdistance — Method

designdistance(C::AbstractBCHCode)

Return the design distance of the BCH code.

CodingTheory.dualdefiningset — Method

dualdefiningset(defset::Vector{Int}, n::Int)

Return the defining set of the dual code of length n and defining set defset.

CodingTheory.expansionfactor — Method

expansionfactor(C::AbstractQuasiCyclicCode)

Return the expansion factor of the quasi-cycle code C.

CodingTheory.generatorpolynomial — Method

generatorpolynomial(C::AbstractCyclicCode)

Return the generator polynomial of the cyclic code.

CodingTheory.generators — Method

generators(C::AbstractQuasiCyclicCode)

Return the generators of the quasi-cyclic code.

CodingTheory.idempotent — Method

idempotent(C::AbstractCyclicCode)

Return the idempotent (polynomial) of the cyclic code.

CodingTheory.index — Method

index(C::AbstractQuasiCyclicCode)

Return the index of the quasi-cyclic code.

CodingTheory.isdegenerate — Method

isdegenerate(C::AbstractCyclicCode)

Return true if the cyclic code is degenerate.

Notes

A cyclic code is degenerate if the parity-check polynomial divides x^r - 1 for

some r less than the length of the code.

CodingTheory.isnarrowsense — Method

isnarrowsense(C::AbstractBCHCode)

Return true if the BCH code is narrowsense.

CodingTheory.isreversible — Method

isreversible(C::AbstractCyclicCode)

Return true if the cyclic code is reversible.

CodingTheory.isselfdual — Method

isselfdual(C::AbstractCyclicCode)

Return whether or not C == dual(C).

CodingTheory.issinglegenerator — Method

issinglegenerator(C::AbstractQuasiCyclicCode)

Return true if C is a single-generator quasi-cyclic code.

CodingTheory.mindistlowerbound — Method

mindistlowerbound(C::AbstractCyclicCode)

Return a lower bound on the minimum distance of the code.

Notes

At the moment, this is only the BCH bound with the Hartmann-Tzeng Bound refinement. The minimum distance is returned if known.

CodingTheory.noncirculantgeneratormatrix — Method

noncirculantgeneratormatrix(C::AbstractQuasiCyclicCode)

Return the non-circulant form of the generator matrix for the quasi-cyclic code C if the polynomial matrix specifies the generator matrix. Otherwise, return missing.

CodingTheory.noncirculantparitycheckmatrix — Method

noncirculantparitycheckmatrix(C::AbstractQuasiCyclicCode)

Return the non-circulant form of the parity-check matrix for the quasi-cyclic code C if the polynomial matrix specifies the parity-check matrix. Otherwise, return missing.

CodingTheory.nonzeros — Method

nonzeros(C::AbstractCyclicCode)

Return the nonzeros of C.

CodingTheory.offset — Method

offset(C::AbstractBCHCode)

Return the offset of the BCH code.

CodingTheory.paritycheckpolynomial — Method

paritycheckpolynomial(C::AbstractCyclicCode)

Return the parity-check polynomial of the cyclic code.

CodingTheory.polynomialmatrix — Method

polynomialmatrix(C::AbstractQuasiCyclicCode)

Return the polynomial matrix used to define the code.

Use polynomialmatrixtype to determine if specifies the generator or parity-check matrix.

CodingTheory.polynomialmatrixtype — Method

polynomialmatrixtype(C::AbstractQuasiCyclicCode)

Return 'G' if the polynomial matrix of C specifies the generator or parity-check matrix.

CodingTheory.polynomialring — Method

polynomialring(C::AbstractCyclicCode)

Return the polynomial ring of the generator polynomial.

CodingTheory.primitiveroot — Method

primitiveroot(C::AbstractCyclicCode)

Return the primitive root of the splitting field.

CodingTheory.qcosets — Method

qcosets(C::AbstractCyclicCode)

Return the q-cyclotomic cosets of the cyclic code.

CodingTheory.qcosetsreps — Method

qcosetsreps(C::AbstractCyclicCode)

Return the set of representatives for the q-cyclotomic cosets of the cyclic code.

CodingTheory.splittingfield — Method

splittingfield(C::AbstractCyclicCode)

Return the splitting field of the generator polynomial.

CodingTheory.type — Method

type(C::AbstractQuasiCyclicCode)

Return the type of the quasi-cycle code C.

CodingTheory.weightmatrix — Method

basematrix(A::MatElem{T}) where T <: ResElem
protographmatrix(A::MatElem{T}) where T <: ResElem
weightmatrix(A::MatElem{T}) where T <: ResElem

Return the base/protograph/weight matrix of A.

Hecke.isprimitive — Method

isprimitive(C::AbstractBCHCode)

Return true if the BCH code is primitive.

Oscar.complement — Method

complement(C::AbstractCyclicCode)

Return the cyclic code whose cyclotomic cosets are the completement of C's.