Linear Codes

CodingTheory.LinearCode — Type

LinearCode(V::AbstractAlgebra.Generic.FreeModule{fq_nmod}, parity::Bool=false)

Return the linear code constructed from the basis of the vector space V. If the optional paramater parity is set to true, a linear code is built with G as the parity-check matrix.

CodingTheory.LinearCode — Type

LinearCode(G::fq_nmod_mat, parity::Bool=false)

Return the linear code constructed with generator matrix G.

Notes

If G is not full rank, a row-reduced form is computed for the generator matrix. The dimension of the code is the number of rows of the full-rank matrix, and the length the number of columns. If the optional paramater parity is set to true, a linear code is built with G as the parity-check matrix.

At the moment, no convention is used for G = 0 and an error is thrown.

Zero columns are not removed.

Examples

julia> F = GF(2);
julia> G = matrix(F, 4, 7, [1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1]);
julia> C = LinearCode(G)
[7, 4]_2 linear code.
Generator matrix: 4 × 7
        1 0 0 0 0 1 1
        0 1 0 0 1 0 1
        0 0 1 0 1 1 0
        0 0 0 1 1 1 1

AbstractAlgebra.VectorSpace — Method

VectorSpace(C::AbstractLinearCode)

Return the code C as a vector space object.

Base.:⊆ — Method

⊆(C1::AbstractLinearCode, C2::AbstractLinearCode)
⊂(C1::AbstractLinearCode, C2::AbstractLinearCode)
issubcode(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return whether or not C1 is a subcode of C2.

Base.in — Method

in(v::Union{fq_nmod_mat, Vector{Int}}, C::AbstractLinearCode)

Return whether or not v is a codeword of C.

Base.iseven — Method

iseven(C::AbstractLinearCode)

Return true if C is even.

Base.length — Method

length(C::AbstractLinearCode)

Return the length of the code.

CodingTheory.Hermitiandual — Method

Hermitiandual(C::AbstractLinearCode)

Return the Hermitian dual of a code defined over a quadratic extension.

CodingTheory.Hermitianhull — Method

Hermitianhull::AbstractLinearCode)

Return the Hermitian hull of C and its dimension.

Notes

The Hermitian hull of a code is the intersection of it and its Hermitian dual.

CodingTheory.areequivalent — Method

areequivalent(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return true if C1 ⊆ C2 and C2 ⊆ C1.

CodingTheory.cardinality — Method

cardinality(C::AbstractLinearCode)

Return the cardinality of the code.

CodingTheory.characteristicpolynomial — Method

characteristicpolynomial(C::AbstractLinearCode)

Return the characteristic polynomial of C.

Notes

The characteristic polynomial is defined in [Lin1999]_

CodingTheory.encode — Method

encode(C::AbstractLinearCode, v::Union{CTMatrixTypes, Vector{Int}})

Return v * G, where G is the generator matrix of C.

CodingTheory.field — Method

field(C::AbstractLinearCode)

Return the base ring of the generator matrix.

CodingTheory.generatormatrix — Function

generatormatrix(C::AbstractLinearCode, standform::Bool=false)

Return the generator matrix of the code.

Notes

If the optional parameter standform is set to true, the standard form of the generator matrix is returned instead.

CodingTheory.genus — Method

genus(C::AbstractLinearCode)

Return the genus, n + 1 - k - d, of the code.

CodingTheory.hull — Method

hull(C::AbstractLinearCode)
Euclideanhull(C::AbstractLinearCode)

Return the (Euclidean) hull of C and its dimension.

Notes

The hull of a code is the intersection of it and its dual.

CodingTheory.isHermitianLCD — Method

isHermitianLCD(C::AbstractLinearCode)

Return true if C is linear complementary Hermitian dual.

Notes

A code is linear complementary Hermitian dual if the dimension of Hermitianhull(C) is zero.

CodingTheory.isHermitiandualcontaining — Method

isHermitiandualcontaining(C::AbstractLinearCode)

Return true if Hermitiandual(C) ⊆ C.

CodingTheory.isHermitianselfdual — Method

isHermitianselfdual(C::AbstractLinearCode)

Return true if areequivalent(C, Hermitiandual(C)).

CodingTheory.isHermitianselforthogonal — Method

isHermitianselforthogonal(C::AbstractLinearCode)
isHermitianweaklyselfdual(C::AbstractLinearCode)

Return true if C ⊆ Hermitiandual(C).

CodingTheory.isLCD — Method

isLCD(C::AbstractLinearCode)

Return true if C is linear complementary dual.

Notes

A code is linear complementary dual if the dimension of hull(C) is zero.

CodingTheory.isMDS — Method

isMDS(C::AbstractLinearCode)

Return true if code is maximum distance separable (MDS).

CodingTheory.isdoublyeven — Method

isdoublyeven(C::AbstractLinearCode)

Return true if C is doubly-even.

CodingTheory.isdualcontaining — Method

isdualcontaining(C::AbstractLinearCode)
isEuclideandualcontaining(C::AbstractLinearCode)

Return true if dual(C) ⊆ C.

CodingTheory.isselfdual — Method

isselfdual(C::AbstractLinearCode)

Return true if areequivalent(C, dual(C)).

CodingTheory.isselforthogonal — Method

isselforthogonal(C::AbstractLinearCode)
isweaklyselfdual(C::AbstractLinearCode)
isEuclideanselforthogonal(C::AbstractLinearCode)

Return true if C ⊆ dual(C).

CodingTheory.istriplyeven — Method

istriplyeven(C::AbstractLinearCode)

Return true if C is triply-even.

CodingTheory.minimumdistancelowerbound — Method

minimumdistancelowerbound(C::AbstractLinearCode)

Return the current lower bound on the minimum distance of C.

CodingTheory.minimumdistanceupperbound — Method

minimumdistanceupperbound(C::AbstractLinearCode)

Return the current upper bound on the minimum distance of C.

CodingTheory.numbercorrectableerrors — Method

numbercorrectableerrors(C::AbstractLinearCode)

Return the number of correctable errors for the code.

Notes

The number of correctable errors is t = floor((d - 1) / 2).

CodingTheory.paritycheckmatrix — Function

paritycheckmatrix(C::AbstractLinearCode, standform::Bool=false)

Return the parity-check matrix of the code.

Notes

If the optional parameter standform is set to true, the standard form of the parity-check matrix is returned instead.

CodingTheory.rate — Method

rate(C::AbstractLinearCode)

Return the rate, R = k/n, of the code.

CodingTheory.relativedistance — Method

relativedistance(C::AbstractLinearCode)

Return the relative minimum distance, δ = d / n of the code if d is known, otherwise return missing.

CodingTheory.setdistancelowerbound! — Method

setdistancelowerbound!(C::AbstractLinearCode, l::Int)

Set the lower bound on the minimum distance of C, if l is better than the current bound.

CodingTheory.setdistanceupperbound! — Method

setdistanceupperbound!(C::AbstractLinearCode, u::Int)

Set the upper bound on the minimum distance of C, if u is better than the current bound.

CodingTheory.setminimumdistance! — Method

setminimumdistance(C::AbstractLinearCode, d::Int)

Set the minimum distance of the code to d.

Notes

The only check done on the value of d is that 1 ≤ d ≤ n.

CodingTheory.syndrome — Method

syndrome(C::AbstractLinearCode, v::Union{fq_nmod_mat, Vector{Int}})

Return Hv, where H is the parity-check matrix of C.

CodingTheory.words — Function

words(C::AbstractLinearCode, onlyprint::Bool=false)
codewords(C::AbstractLinearCode, onlyprint::Bool=false)
elements(C::AbstractLinearCode, onlyprint::Bool=false)

Return the elements of C.

Notes

If onlyprint is true, the elements are only printed to the console and not returned.

Hecke.dimension — Method

dimension(C::AbstractLinearCode)

Return the dimension of the code.

Hecke.dual — Method

dual(C::AbstractLinearCode)
Euclideandual(C::AbstractLinearCode)

Return the (Euclidean) dual of the code C.