Modifying Codes

CodingTheory.:⊕ — Method

⊕(C1::AbstractLinearCode, C2::AbstractLinearCode)
directsum(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊕ C2

Return the direct sum code of C1 and C2.

Notes

The direct sum code has generator matrix G1 ⊕ G2 and parity-check matrix H1 ⊕ H2.

CodingTheory.:⊗ — Method

⊗(C1::AbstractLinearCode, C2::AbstractLinearCode)
kron(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
tensorproduct(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
directproduct(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
productcode(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2

Return the (direct/tensor) product code of C1 and C2.

Notes

The product code has generator matrix G1 ⊗ G2.

CodingTheory.augment — Method

augment(C::AbstractLinearCode, M::fq_nmod_mat)

Return the code of C whose generator matrix is augmented with M.

Notes

Vertically joins the matrix M to the bottom of the generator matrix of C.

CodingTheory.codecomplement — Method

codecomplement(C1::AbstractLinearCode, C2::AbstractLinearCode)
quo(C1::AbstractLinearCode, C2::AbstractLinearCode)
quotient(C1::AbstractLinearCode, C2::AbstractLinearCode)
/(C2::AbstractLinearCode, C1::AbstractLinearCode)

Return the code C2 / C1 given C1 ⊆ C2.

CodingTheory.constructionX — Method

constructionX(C1::AbstractLinearCode, C2::AbstractLinearCode, C3::AbstractLinearCode)

Return the code generated by the construction X procedure.

Notes

Let C1 be an [n, k, d], C2 be an [n, k - l, d + e], and C3 be an [m, l, e] linear code with C2 ⊂ C1 be proper. Construction X creates a [n + m, k, d + e] linear code.

CodingTheory.constructionX3 — Method

constructionX3(C1::AbstractLinearCode, C2::AbstractLinearCode, C3::AbstractLinearCode,
    C4::AbstractLinearCode, C5::AbstractLinearCode))

Return the code generated by the construction X3 procedure.

Notes

Let C1 = [n, k1, d1], C2 = [n, k2, d2], C3 = [n, k3, d3], C4 = [n4, k2 - k1, d4], and C5 = [n5, k3 - k2, d5] with C1 ⊂ C2 ⊂ C3. Construction X3 creates an [n + n4 + n5, k3, d] linear code with d ≥ min{d1, d2 + d4, d3 + d5}.

CodingTheory.expandedcode — Method

expandedcode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the expanded code of C constructed by exapnding the generator matrix to the subfield K using the provided dual basis for the field of C over K.

CodingTheory.expurgate — Method

expurgate(C::AbstractLinearCode, rows::Vector{Int})
expurgate(C::AbstractLinearCode, rows::Int)

Return the code of C expuragated at the rows in rows.

Notes

Deletes the rows from the generator matrix and then removes any potentially resulting zero columns.

CodingTheory.juxtaposition — Method

juxtaposition(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the code generated by the horizontal concatenation of the generator matrices of C1 then C2.

CodingTheory.lengthen — Method

lengthen(C::AbstractLinearCode)

Return the lengthened code of C.

Notes

This augments the all 1's row and then extends.

CodingTheory.permutecode — Method

permutecode(C::AbstractLinearCode, σ::Union{PermGroupElem, Perm{T}, Vector{T}}) where T <: Int

Return the code whose generator matrix is C's with the columns permuted by σ.

Notes

If σ is a vector, it is interpreted as the desired column order for the generator matrix of C.

CodingTheory.puncture — Method

puncture(C::AbstractLinearCode, cols::Vector{Int})
puncture(C::AbstractLinearCode, cols::Int)

Return the code of C punctured at the columns in cols.

Notes

Deletes the columns from the generator matrix and then removes any potentially resulting zero rows.

CodingTheory.subcode — Method

subcode(C::AbstractLinearCode, k::Int)

Return a k-dimensional subcode of C.

CodingTheory.subcode — Method

subcode(C::AbstractLinearCode, rows::Vector{Int})

Return a subcode of C using the rows of the generator matrix of C listed in rows.

CodingTheory.subcodeofdimensionbetweencodes — Method

subcodeofdimensionbetweencodes(C1::AbstractLinearCode, C2::AbstractLinearCode, k::Int)

Return a subcode of dimenion k between C1 and C2.

Notes

This function arguments generators of C1 / C2 to C2 until the desired dimenion is reached.

CodingTheory.subfieldsubcode — Method

subfieldsubcode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the subfield subcode code of C over K using the provided dual basis for the field of C over K.

CodingTheory.tracecode — Method

tracecode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the trace code of C over K using the provided dual basis for the field of C over K using Delsarte's theorem.

CodingTheory.upluswvpluswuplusvplusw — Method

upluswvpluswuplusvplusw(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the code generated by the (u + w | v + w | u + v + w)-construction.

Notes

Let C1 = [n, k1, d1] and C2 = [n, k2, d2]. This construction produces an [3n, 2k1 + k2] linear code. For binary codes, wt(u + w | v + w | u + v + w) = 2 wt(u ⊻ v) - wt(w) + 4s, where s = |{i | ui = vi = 0, w_i = 1}|.

CodingTheory.uuplusv — Method

uuplusv(C1::AbstractLinearCode, C2::AbstractLinearCode)
Plotkinconstruction(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the Plotkin (u | u + v)-construction with u ∈ C1 and v ∈ C2.