Product Codes

Constructors

Attributes

Methods

CodingTheory.Hermitiandual — Method

Hermitiandual(C::MatrixProductCode)

Return the Hermitian dual of C. If the dual is also a matrix product code, a matrix product code will be returned.

Hecke.dual — Method

dual(C::MatrixProductCode)

Return the dual of C. If the dual is also a matrix product code, a matrix product code will be returned.