Quasi-Cyclic Codes

Quasi-cyclic codes are a subtype of LinearCode and inherit its methods. While quasi-cyclic codes may be seen as generalizations of cyclic codes, here they are treated as independent topics.

Constructors

CodingTheory.QuasiCyclicCode — Type

QuasiCyclicCode(A::MatElem{T}, parity::Bool=false) where T <: ResElem

Return the quasi-cycle code specified by the matrix A of polynomial circulant generators. If the optional paramater parity is set to true, the input is used to construct the parity-check matrix.

QuasiCyclicCode(v::Vector{fq_nmod_mat}, l::Int, circgens::Bool, parity::Bool=false)

Return the quasi-cyclic code of index l generated by right-bit shifts of size l of the generator vectors v. If circgens is true, the vectors are taken to be (column) generators for the circulant matrices instead of generator vectors for the code. If the optional paramater parity is set to true, the input is used to construct the parity-check matrix.

Notes

If circgens is false, then the length of the code is ncols(v[1]) and must be divisible by l.

If circgens is true, then the length of the code is ncols(v[1]) * l. Circulant matrices are stacked in rows of length l, so l must divide length(v).

QuasiCyclicCode(v::fq_nmod_mat, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l generated by right-bit shifts of size l of the generator vector v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

QuasiCyclicCode(v::Vector{fq_nmod_poly}, n::Int, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l whose circulants are defined by the generator polynomials v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

QuasiCyclicCode(v::Vector{AbstractCyclicCode}, l::Int, parity::Bool=false)

Return the quasi-cyclic code of index l whose circulants are determined by the cyclic codes in v. If the optional paramater parity is set to true, the input is used to construct the parity check matrix.

Attributes

CodingTheory.index — Function

index(C::AbstractQuasiCyclicCode)

Return the index of the quasi-cyclic code.

CodingTheory.expansionfactor — Function

expansionfactor(C::AbstractQuasiCyclicCode)

Return the expansion factor of the quasi-cycle code C.

CodingTheory.issinglegenerator — Function

issinglegenerator(C::AbstractQuasiCyclicCode)

Return true if C is a single-generator quasi-cyclic code.

CodingTheory.polynomialmatrix — Function

polynomialmatrix(C::AbstractQuasiCyclicCode)

Return the polynomial matrix used to define the code.

Use polynomialmatrixtype to determine if specifies the generator or parity-check matrix.

CodingTheory.polynomialmatrixtype — Function

polynomialmatrixtype(C::AbstractQuasiCyclicCode)

Return 'G' if the polynomial matrix of C specifies the generator or parity-check matrix.

The type parameter is either :G or :H, specifying whether the polynomial matrix represents the generator or parity-check matrix.

CodingTheory.type — Function

type(C::AbstractQuasiCyclicCode)

Return the type of the quasi-cycle code C.

type(W::WeightEnumerator)

Returns the type of the weight enumerator W.

Methods

The following are not computed and stored at the time of construction and must be computed by using these methods.

CodingTheory.weightmatrix — Function

basematrix(A::MatElem{T}) where T <: ResElem
protographmatrix(A::MatElem{T}) where T <: ResElem
weightmatrix(A::MatElem{T}) where T <: ResElem

Return the base/protograph/weight matrix of A.

Missing docstring.

Missing docstring for generatormatrix. Check Documenter's build log for details.

Missing docstring.

Missing docstring for paritycheckmatrix. Check Documenter's build log for details.

CodingTheory.noncirculantgeneratormatrix — Function

noncirculantgeneratormatrix(C::AbstractQuasiCyclicCode)

Return the non-circulant form of the generator matrix for the quasi-cyclic code C if the polynomial matrix specifies the generator matrix; otherwise, return missing.

CodingTheory.noncirculantparitycheckmatrix — Function

noncirculantparitycheckmatrix(C::AbstractQuasiCyclicCode)

Return the non-circulant form of the parity-check matrix for the quasi-cyclic code C if the polynomial matrix specifies the parity-check matrix; otherwise, return missing.

CodingTheory.generators — Function

generators(C::AbstractQuasiCyclicCode)

Return the generators of the quasi-cyclic code.

CodingTheory.circulants — Function

circulants(C::AbstractQuasiCyclicCode)

Return the circulant matrices of the quasi-cyclic code.