Miscellaneous Known Stabilizer Codes

CodingTheory.BaconShorCode — Method

BaconShorCode(d::Int)

Return the Bacon-Shor subsystem code on a d x d lattice.

CodingTheory.HCode — Method

HCode(k::Int)

Return the [[k + 4, k, 2]] H code from https://errorcorrectionzoo.org/c/quantum_h.

CodingTheory.PlanarSurfaceCode — Method

PlanarSurfaceCode(dx::Int, dz::Int)
PlanarSurfaceCode(d::Int)

Return the [[dx * dz + (dx - 1) * (dz - 1), 1, dx/dz]] planar surface code.

The top and bottom boundaries are "smooth" (Z) and the left and right are "rough" (X).

CodingTheory.Q1573 — Method

Q1573()

Return the [[15, 7, 3]] quantum Hamming code.

CodingTheory.Q15RM — Method

Q15RM()
Q1513()

Return the [[15, 1, 3]] quantum Reed-Muller code.

CodingTheory.ShorCode — Method

Shorcode()
Q913()

Return the [[9, 1, 3]] Shor code.

CodingTheory.SteaneCode — Method

Steanecode()
Q713()

Return the [[7, 1, 3]] Steane code.

CodingTheory.ToricCode — Method

ToricCode(d::Int)

Return the [[2d^2, 2, d]] toric code.

The lattice orientation used here follows the picture at https://errorcorrectionzoo.org/c/surface.

CodingTheory.TriangularColorCode488 — Method

TriangularColorCode488(d::Int)

Return the 4.8.8 triangular color code of distance d with trellis numbering.

CodingTheory.TriangularColorCode666 — Method

TriangularColorCode666(d::Int)

Return the 6.6.6 triangular color code of distance d with trellis numbering.

CodingTheory.XZZXSurfaceCode — Method

XZZXSurfaceCode(d::Int)

Return the [[d^2, 1, d]] XZZX surface code.