Product Codes

Notes

The hypergraph product is defined in "J. Tillich, G. Zémor. Quantum LDPC codes with positive rate and minimum distance proportional to n^(1/2). (2013) arXiv:0903.0566v2"

The generalized Shor code is defined in "D. Bacon and A. Casaccino. Quantum error correcting subsystem codes from two classical linear codes. (2006) http://arxiv.org/abs/quant-ph/0610088"

Hyperbicycle codes are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate" and "Quantum Kronecker sum-product low-density parity-check codes with finite rate".

Generealized bicycle codes are are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate", "Quantum kronecker sum-product low-density parity- check codes with finite rate", and "Degenerate Quantum LDPC Codes With Good Finite Length Performance".

The generalized hypergraph product is defined in "Degenerate Quantum LDPC Codes With Good Finite Length Performance".

The bias-tailored lifted product is defined in "Bias-tailored quantum LDPC codes".

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(A::CTMatrixTypes, B::CTMatrixTypes, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)

Return the hypergraph product code of matrices A and B whose signs are determined by charvec.

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the hypergraph product code of C1 and C2 whose signs are determined by charvec.

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the (symmetric) hypergraph product code of C whose signs are determined by charvec.

CodingTheory.BiasTailoredQuasiCyclicLiftedProductCode — Method

BiasTailoredQuasiCyclicLiftedProductCodeMatrices(A::MatElem{T}, B::MatElem{T}, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm) where T <: ResElem

Return the bias-tailored lifted product code of A and B whose signs are given by charvec.

Arguments

A - an m x n1 matrix with coefficents in F_2[x]/(x^m - 1)

B - an m x n2 matrix with coefficents in the same residue ring

CodingTheory.BiasTailoredQuasiCyclicLiftedProductCodeMatrices — Method

BiasTailoredQuasiCyclicLiftedProductCodeMatrices(A::MatElem{T}, B::MatElem{T}) where T <: ResElem

Return the pre-lifted stabilizer matrix for bias-tailored lifted product code of A and B.

Arguments

A - an m x n1 matrix with coefficents in F_2[x]/(x^m - 1)

B - an m x n2 matrix with coefficents in the same residue ring

Notes

Use BiasTailoredQuasiCyclicLiftedProductCode to return a quantum code over the base ring directly.

CodingTheory.GeneralizedBicycleCode — Method

GeneralizedBicycleCode(A::fq_nmod_mat, B::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)

Return the generealized bicycle code given by A and B whose signs are determined by charvec.

CodingTheory.GeneralizedBicycleCode — Method

GeneralizedBicycleCode(a::T, b::T, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm) where T <: ResElem

Return the generealized bicycle code determined by a and b whose signs are determined by charvec.

Notes

l x l circulant matrices are constructed using the coefficients of the polynomials a and b in F_q[x]/(x^l - 1) (gcd(q, l) = 1) as the first column

CodingTheory.GeneralizedHypergraphProductCodeMatrices — Method

GeneralizedHypergraphProductCodeMatrices(A::MatElem{T}, b::T) where T <: ResElem

Return the pre-lifted matrices HX and HZ of the generalized hypergraph product code of A and b.

Arguments

A - an m x n matrix with coefficents in F_2[x]/(x^m - 1)

b - a polynomial over the same residue ring

Notes

Use LiftedGeneralizedHypergraphProductCode to return a quantum code over the base ring directly.

CodingTheory.GeneralizedShorCode — Function

GeneralizedShorCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)
BaconCasaccinoConstruction(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)

Return the generalized Shor code of C1 and C2 with C1⟂ ⊆ C2 whose signs are determined by charvec.

CodingTheory.HyperBicycleCode — Method

HyperBicycleCode(a::Vector{fq_nmod_mat}, b::Vector{fq_nmod_mat}, χ::Int, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)

Return the hyperbicycle CSS code of a and b given χ whose signs are determined by charvec.

Arguments

a: A vector of length c of binary matrices of the same dimensions.

b: A vector of length c of binary matrices of the same dimensions, potentially different from those of a.

χ: A strictly positive integer coprime with c.

CodingTheory.HyperBicycleCodeCSS — Method

HyperBicycleCodeCSS(a::Vector{fq_nmod_mat}, b::Vector{fq_nmod_mat}, χ::Int, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm)

Return the hyperbicycle CSS code of a and b given χ whose signs are determined by charvec.

Arguments

a: A vector of length c of binary matrices of the same dimensions.

b: A vector of length c of binary matrices of the same dimensions, potentially different from those of a.

χ: A strictly positive integer coprime with c.

CodingTheory.LiftedGeneralizedHypergraphProductCode — Method

LiftedGeneralizedHypergraphProductCode(A::MatElem{T}, b::T, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm) where T <: ResElem

Return the lifted generalized hypergraph product code of A and b.

Arguments

A - an m x n matrix with coefficents in F_2[x]/(x^m - 1)

b - a polynomial over the same residue ring

CodingTheory.QuasiCyclicLiftedProductCode — Method

QuasiCyclicLiftedProductCode(A::MatElem{T}, B::MatElem{T}, charvec::Union{Vector{nmod}, Missing}=missing, logsalg::Symbol=:stndfrm) where T <: ResElem

Return the quasi-cyclic lifted product code given by the matrices A and B and whose signs are determined by charvec.

CodingTheory.QuasiCyclicLiftedProductCodeMatrices — Method

QuasiCyclicLiftedProductCodeMatrices(A::MatElem{T}, B::MatElem{T}) where T <: ResElem

Return the pre-lifted matrices HX and HZ for the lifted quasi-cyclic lifted product code.

Arguments

A - an m x n1 matrix with coefficents in F_2[x]/(x^m - 1)

B - an m x n2 matrix with coefficents in the same residue ring

Notes

Use QuasiCyclicLiftedProductCode to return a quantum code over the base ring directly.