Reed-Muller Codes

CodingTheory.ReedMullerMod.ReedMullerCode — Type

ReedMullerCode(q::Integer, r::Integer, m::Integer, verify::Bool=true)

Return the RM(r, m) Reed-Muller code over GF(q).

If the optional parameter verify is set to true, basic checks are done to ensure correctness.

CodingTheory.ReedMullerMod.RMm — Method

RMm(C::ReedMullerCode)

Return the parameter m of the RM(r, m) Reed-Muller code.

CodingTheory.ReedMullerMod.ReedMullergeneratormatrix — Method

ReedMullergeneratormatrix(q::Integer, r::Integer, m::Integer)

Return the recursive form of the generator matrix for the RM(r, m) Reed-Muller code over GF(q).

CodingTheory.ReedMullerMod.dual — Method

dual(C::ReedMullerCode)

Return the dual of the Reed-Muller code C.

CodingTheory.ReedMullerMod.order — Method

order(C::ReedMullerCode)
RMr(C::ReedMullerCode)

Return the order, r, of the RM(r, m) Reed-Muller code.