Hypergraph Product Codes

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the hypergraph product code of C1 and C2.

The hypergraph product is defined in "J. Tillich, G. Zémor. Quantum LDPC codes with positive rate and minimum distance proportional to n^(1/2). (2013) arXiv:0903.0566v2"

Arguments

C1: a binary linear code

C2: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the (symmetric) hypergraph product code of C.

The hypergraph product is defined in "J. Tillich, G. Zémor. Quantum LDPC codes with positive rate and minimum distance proportional to n^(1/2). (2013) arXiv:0903.0566v2"

Arguments

C: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.GeneralizedShorCode — Function

GeneralizedShorCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)
BaconCasaccinoConstruction(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the generalized Shor code of C1 and C2 with C1⟂ ⊆ C2.

The generalized Shor code is defined in "D. Bacon and A. Casaccino. Quantum error correcting subsystem codes from two classical linear codes. (2006) http://arxiv.org/abs/quant-ph/0610088"

Arguments

C1: a binary linear code

C2: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.