Miscellaneous Known Codes

CodingTheory.ExtendedGolayCode — Method

ExtendedGolayCode(p::Int)

Return the [24, 12, 8] extended binary Golay code if p == 2 or the [12, 6, 6] extended ternary Golay code if p == 3.

The [23, 12, 7] binary Golay code may be obtained by puncturing this code in any coordinate.

Notes

This is constructed using an explicit matrix. The Golay codes are constructed

by calling puncture(ExtendedGolayCode(p), 1). All single punctures are equivalent.

CodingTheory.GolayCode — Method

GolayCode(p::Int)

Return the [23, 12, 7]binary Golay code ifp == 2or the[11, 6, 5]ternary Golay code ifp == 3`.

The [24, 12, 8] extended binary Golay code may be obtained by extending this code. Adding an overall parity check to the ternary Golay code will give either a [12, 6, 6] if punctured and extended in the first coordinate or a [12, 6, 5] code otherwise.

Notes

These codes are constructed by calling puncture(ExtendedGolayGode(p), [1]).

All single punctures are equivalent.

CodingTheory.HammingCode — Method

HammingCode(q::Int, r::Int)

Return the [(q^r - 1)/(q - 1), (q^r - 1)/(q - 1) - r, 3] Hamming code over GF(q).

Notes

This is currently only implemented for binary codes.

CodingTheory.Hexacode — Method

Hexacode()

Return the [6, 3, 4] hexacode over GF(4).

CodingTheory.RepetitionCode — Method

RepetitionCode(q::Int, n::Int)

Return the [n, 1, n] repetition code over GF(q).

CodingTheory.SimplexCode — Method

SimplexCode(q::Int, r::Int)

Return the [(q^r - 1)/(q - 1), r] simplex code over GF(q).

Generator matrices for the binary codes are constructed using the standard recursive definition. The higher fields return dual(HammingCode(q, r)).

Notes

This is currently only implemented for binary codes.

CodingTheory.TetraCode — Method

TetraCode()

Return the [4, 2, 3] tetra code over GF(3).

This is equiavlent to the Hamming(3, 2, 3) code, but the construction here is based on the commonly presented generator and parity-check matrices.