Product Codes

CodingTheory.Hermitiandual — Method

Hermitiandual(C::MatrixProductCode)

Return the Hermitian dual of C. If the dual is also a matrix product code, a matrix product code will be returned.

Hecke.dual — Method

dual(C::MatrixProductCode)

Return the dual of C. If the dual is also a matrix product code, a matrix product code will be returned.

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the hypergraph product code of C1 and C2.

The hypergraph product is defined in "J. Tillich, G. Zémor. Quantum LDPC codes with positive rate and minimum distance proportional to n^(1/2). (2013) arXiv:0903.0566v2"

Arguments

C1: a binary linear code

C2: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.HypergraphProductCode — Type

HypergraphProductCode(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the (symmetric) hypergraph product code of C.

The hypergraph product is defined in "J. Tillich, G. Zémor. Quantum LDPC codes with positive rate and minimum distance proportional to n^(1/2). (2013) arXiv:0903.0566v2"

Arguments

C: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.BiasTailoredQuasiCyclicLiftedProductCode — Method

BiasTailoredQuasiCyclicLiftedProductCode(A::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.Res{fq_nmod_poly}},

B::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.Res{fqnmodpoly}})

Return the bias-tailored lifted product code of A and B with entries over the residue ring F_2[x]/(x^m - 1).

The bias-tailored lifted product is defined in Bias-tailored quantum LDPC codes.

CodingTheory.GeneralizedBicycleCode — Method

GeneralizedBicycleCode(a::AbstractAlgebra.Generic.Res{fq_nmod_poly}, b::AbstractAlgebra.Generic.Res{fq_nmod_poly})

Return the generealized bicycle code determined by a and b.

l x l circulant matrices are constructed using the coefficients of the polynomials a and b in F_q[x]/(x^l - 1) (gcd(q, l) = 1) as the first column.

Generealized bicycle codes are are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate", "Quantum kronecker sum-product low-density parity- check codes with finite rate", and "Degenerate Quantum LDPC Codes With Good Finite Length Performance".

CodingTheory.GeneralizedBicycleCode — Method

GeneralizedBicycleCode(A::fq_nmod_mat, B::fq_nmod_mat)

Return the generealized bicycle code given by A and B.

Generealized bicycle codes are are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate", "Quantum kronecker sum-product low-density parity- check codes with finite rate", and "Degenerate Quantum LDPC Codes With Good Finite Length Performance".

CodingTheory.GeneralizedHypergraphProductCode — Method

GeneralizedHypergraphProductCode(A::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.Res{fq_nmod_poly}}, b::AbstractAlgebra.Generic.Res{fq_nmod_poly})

Return the two matrices HX and HZ of generalized hypergraph product of A and b over the reside ring.

The generalized hypergraph product is defined in "Degenerate Quantum LDPC Codes With Good Finite Length Performance". To return a quantum code directly, use LiftedGeneralizedHypergraphProductCode.

Notes

A- anm x n`matrix with coefficents in a residue ring overGF(2)`.

b - a polynomial over the same residue ring

CodingTheory.GeneralizedShorCode — Function

GeneralizedShorCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)
BaconCasaccinoConstruction(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the generalized Shor code of C1 and C2 with C1⟂ ⊆ C2.

The generalized Shor code is defined in "D. Bacon and A. Casaccino. Quantum error correcting subsystem codes from two classical linear codes. (2006) http://arxiv.org/abs/quant-ph/0610088"

Arguments

C1: a binary linear code

C2: a binary linear code

charvec: a length 2n vector with elements in the Z/(4). The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.HyperBicycleCode — Method

HyperBicycleCode(a::Vector{fq_nmod_mat}, b::Vector{fq_nmod_mat}, χ::Int)

Return the hyperbicycle CSS code of a and b given χ.

Hyperbicycle codes are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate" and "Quantum Kronecker sum-product low-density parity-check codes with finite rate".

Inputs:

a: A vector of length c of binary fq_nmod_mat matrices of the same dimensions.

b: A vector of length c of binary fq_nmod_mat matrices of the same dimensions, potentially different from those of a.

χ: A strictly positive integer coprime with c.

CodingTheory.HyperBicycleCodeCSS — Method

HyperBicycleCodeCSS(a::Vector{fq_nmod_mat}, b::Vector{fq_nmod_mat}, χ::Int)

Return the hyperbicycle CSS code of a and b given χ.

Hyperbicycle codes are found in "Quantum ``hyperbicycle'' low-density parity check codes with finite rate" and "Quantum Kronecker sum-product low-density parity-check codes with finite rate".

Inputs:

a: A vector of length c of binary fq_nmod_mat matrices of the same dimensions.

b: A vector of length c of binary fq_nmod_mat matrices of the same dimensions, potentially different from those of a.

χ: A strictly positive integer coprime with c.

CodingTheory.LiftedGeneralizedHypergraphProductCode — Method

LiftedGeneralizedHypergraphProductCode(A::AbstractAlgebra.Generic.MatSpaceElem{AbstractAlgebra.Generic.Res{fq_nmod_poly}}, b::AbstractAlgebra.Generic.Res{fq_nmod_poly})

Return the CSS code produced by lifting the generalized hypergraph product of A and b over the underlying base field.