Tilings

CodingTheory.CoxeterMatrix — Type

struct CoxeterMatrix <: AbstractMatrix{Int64}

A symmetric integer Coxeter matrix stored by its upper-triangular entries.

CodingTheory.ReflectionGroup — Type

struct ReflectionGroup

A GAP reflection group together with its generators and defining CoxeterMatrix.

CodingTheory.coset_intersection — Function

coset_intersection(
    gen_idx_A::Vector{Int64},
    gen_idx_B::Vector{Int64},
    subgroup::GAP.GapObj,
    G::ReflectionGroup
) -> SparseArrays.SparseMatrixCSC{Int64, Int64}
coset_intersection(
    gen_idx_A::Vector{Int64},
    gen_idx_B::Vector{Int64},
    subgroup::GAP.GapObj,
    G::ReflectionGroup,
    transversal::Union{Nothing, GAP.GapObj}
) -> SparseArrays.SparseMatrixCSC{Int64, Int64}

Return the intersection of the cosets of g/subgroup wrt gen_idx_A and wrt gen_idx_B.

Notes

This outputs a sparse matrix with rows indexing the gen_idx_A cosets and columns indexing the gen_idx_B cosets.

CodingTheory.cycle_tetrahedron_group — Method

cycle_tetrahedron_group(
    q::Int64,
    r::Int64,
    s::Int64,
    t::Int64
) -> ReflectionGroup

Return the "cycle" Coxeter group with high-order (>2) relations given by q, r, s, and t.

Notes

The corresponding Coxeter diagram is:

   q
 o---o
t|   |r
 o---o
   s

CodingTheory.is_fixed_point_free — Method

is_fixed_point_free(
    subgroup::GAP.GapObj,
    G::ReflectionGroup
) -> Bool

Return true if the subgroup of g is fixed-point free; otherwise false.

CodingTheory.is_k_colorable — Method

is_k_colorable(
    k::Int64,
    gen_idx::AbstractVector{<:Int64},
    translations::AbstractVector{<:GAP.GapObj},
    subgroup::GAP.GapObj,
    G::ReflectionGroup
) -> Any

Return true if the group elements corresponding to gen_idx in g/subgroup are k-colorable; otherwise false.

CodingTheory.is_orientable — Method

is_orientable(
    subgroup::GAP.GapObj,
    G::ReflectionGroup
) -> Any

Return true if subgroup lies in the orientation-preserving subgroup of G generated by products of pairs of reflections; otherwise, return false.

CodingTheory.q_r_s_group — Method

q_r_s_group(q::Int64, r::Int64, s::Int64) -> ReflectionGroup

Return the Coxeter group corresponding to Schläfli symbol {q, r, s}.

Notes

The corresponding Coxeter diagram is:

o---o---o---o
  q   r   s

CodingTheory.r_s_group — Method

r_s_group(r::Int64, s::Int64) -> ReflectionGroup

Return the Coxeter group corresponding to Schläfli symbol {r, s}.

Notes

The corresponding Coxeter diagram is:

o---o---o
  r   s

CodingTheory.simplex_group — Method

simplex_group(cox_mat::CoxeterMatrix) -> ReflectionGroup

Return the reflection group with fundamental simplex specified by cox_mat.

CodingTheory.star_tetrahedron_group — Method

star_tetrahedron_group(
    q::Int64,
    r::Int64,
    s::Int64
) -> ReflectionGroup

Return the "star" Coxeter group with higher-order (>2) relations given by q, r, and s.

Notes

The corresponding Coxeter diagram is:

      o
     / r
o---o
  q  \ s
      o

CodingTheory.tetrahedron_group — Method

tetrahedron_group(orders::Vector{Int64}) -> ReflectionGroup

Return the tetrahedron group with relations given by orders.

CodingTheory.triangle_group — Method

triangle_group(
    l::Int64,
    m::Int64,
    n::Int64
) -> ReflectionGroup

Return the (l, m, n) triangle group.

Oscar.normal_subgroups — Method

normal_subgroups(
    G::ReflectionGroup,
    max_index::Integer
) -> Any

Return all normal subgroups of g with index up to max_index.