Generalized Reed-Solomon Codes

CodingTheory.GeneralizedReedSolomonCode — Method

GeneralizedReedSolomonCode(k::Int, v::Vector{fq_nmod}, γ::Vector{fq_nmod})

Return the dimension k Generalized Reed-Solomon code with scalars v and evaluation points γ.

Notes

The vectors v and γ must have the same length and every element must be over the same field.

The elements of v need not be distinct but must be nonzero.

The elements of γ must be distinct.

CodingTheory.dualscalars — Method

dualscalars(C::GeneralizedReedSolomonCode)

Return the scalars of the dual of the Generalized Reed-Solomon code C.

CodingTheory.evaluationpoints — Method

evaluationpoints(C::GeneralizedReedSolomonCode)

Return the evaluation points γ of the Generalized Reed-Solomon code C.

CodingTheory.scalars — Method

scalars(C::GeneralizedReedSolomonCode)

Return the scalars v of the Generalized Reed-Solomon code C.

Hecke.dual — Method

dual(C::GeneralizedReedSolomonCode)

Return the dual of the Generalized Reed-Solomon code.