Generalized Reed-Solomon Codes

Generalized Reed-Solomon codes are a subtype of LinearCode and inherit its methods.

Constructors

CodingTheory.GeneralizedReedSolomonCode — Type

GeneralizedReedSolomonCode(k::Int, v::Vector{fq_nmod}, γ::Vector{fq_nmod})

Return the dimension k Generalized Reed-Solomon code with scalars v and evaluation points γ.

Notes

The vectors v and γ must have the same length and every element must be over the same field.

The elements of v need not be distinct but must be nonzero.

The elements of γ must be distinct.

Attributes

CodingTheory.scalars — Function

scalars(C::GeneralizedReedSolomonCode)

Return the scalars v of the Generalized Reed-Solomon code C.

CodingTheory.dualscalars — Function

dualscalars(C::GeneralizedReedSolomonCode)

Return the scalars of the dual of the Generalized Reed-Solomon code C.

CodingTheory.evaluationpoints — Function

evaluationpoints(C::GeneralizedReedSolomonCode)

Return the evaluation points γ of the Generalized Reed-Solomon code C.

Methods