Reed-Muller Codes

Reed-Muller codes are a subtype of LinearCode and inherit its methods.

Constructors

The (binary) Reed-Muller family is generated using the recursive, $(u \mid u + v)$-form of the generator matrices. Different sources use different conventions for the base case generator matrix. If alt is true, the identity is used for the generator matrix for $\mathcal{RM}(1, 1)$; otherwise, $\begin{pmatrix} 1 & 1\\ 0 & 1\end{pmatrix}$ is used.

CodingTheory.ReedMullerCode — Type

ReedMullerCode(r::Int, m::Int, alt::Bool=false)

Return the $\mathcal{RM}(r, m)$ Reed-Muller code.

Notes

If alt is true, the identity is used for the generator matrix for $\mathcal{RM}(1, 1)$, as in common in some sources. Otherwise, [1 1; 0 1] is used, as is common in other sources.

Attributes

GroupsCore.order — Function

order(C::ReedMullerCode)
RM_r(C::ReedMullerCode)

Return the order, r, of the $\mathcal{RM}(r, m)$ Reed-Muller code.

CodingTheory.number_of_variables — Function

number_of_variables(C::ReedMullerCode)
RM_m(C::ReedMullerCode)

Return the number of variables, m, of the $\mathcal{RM}(r, m)$ Reed-Muller code.

Methods