Cyclic Codes

Cyclic codes are a subtype of LinearCode and inherit its methods. For more information on how to use these functions, see the cyclic code tutorial.

Cyclotomic Cosets

The following set of functions are useful for defining cyclic codes.

CodingTheory.ord — Function

ord(n::Int, q::Int)

Return the order of n mod q.

CodingTheory.cyclotomiccoset — Function

cyclotomiccoset(x::Int, q::Int, n::Int, tosort::Bool=true, verbose::Bool=false)

Return the q-cyclotomic coset of x modulo n.

Notes

If the optional parameter tosort is set to false, the result will not be

sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.allcyclotomiccosets — Function

allcyclotomiccosets(q::Int, n::Int, tosort::Bool=true, verbose::Bool=false)

Return all q-cyclotomic cosets modulo n.

Notes

If the optional parameter tosort is set to false, the result will not be

sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.complementqcosets — Function

complementqcosets(q::Int, n::Int, qcosets::Vector{Vector{Int64}})

Return the complement of the q-cyclotomic cosets modulo n of qcosets.

CodingTheory.qcosetpairings — Function

qcosetpairings(arr::Vector{Vector{Int64}}, n::Int)

Return the q-cyclotomic cosets modulo n collected into complementary pairs.

Missing docstring.

Missing docstring for qcosettable. Check Documenter's build log for details.

CodingTheory.dualqcosets — Function

dualqcosets(q::Int, n::Int, qcosets::Vector{Vector{Int64}})

Return the dual of the q-cyclotomic cosets modulo n of qcosets.

Constructors

CodingTheory.CyclicCode — Type

CyclicCode(q::Int, n::Int, cosets::Vector{Vector{Int}})

Return the CyclicCode of length n over GF(q) with q-cyclotomic cosets cosets.

Notes

This function will auto determine if the constructed code is BCH or Reed-Solomon

and call the appropriate constructor.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> cosets = definingset([i for i = b:(b + δ - 2)], q, n, false);
julia> C = CyclicCode(q, n, cosets)

CyclicCode(n::Int, g::fq_nmod_poly)

Return the length n cyclic code generated by the polynomial g.

CodingTheory.BCHCode — Type

BCHCode(q::Int, n::Int, δ::Int, b::Int=0)

Return the BCHCode of length n over GF(q) with design distance δ and offset b.

Notes

This function will auto determine if the constructed code is Reed-Solomon

and call the appropriate constructor.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> B = BCHCode(q, n, δ, b)
[15, 5, ≥7; 1]_2 BCH code over splitting field GF(16).
2-Cyclotomic cosets:
        C_1 ∪ C_3 ∪ C_5
Generator polynomial:
        x^10 + x^8 + x^5 + x^4 + x^2 + x + 1
Generator matrix: 5 × 15
        1 1 1 0 1 1 0 0 1 0 1 0 0 0 0
        0 1 1 1 0 1 1 0 0 1 0 1 0 0 0

        0 0 1 1 1 0 1 1 0 0 1 0 1 0 0
        0 0 0 1 1 1 0 1 1 0 0 1 0 1 0
        0 0 0 0 1 1 1 0 1 1 0 0 1 0 1

BCHCode(C::AbstractCyclicCode)

Return the BCH supercode of the cyclic code C.

CodingTheory.ReedSolomonCode — Type

ReedSolomonCode(q::Int, δ::Int, b::Int=0)

Return the ReedSolomonCode over GF(q) with distance d and offset b.

Examples

julia> ReedSolomonCode(8, 3, 0)
[7, 5, ≥3; 0]_8 Reed Solomon code.
8-Cyclotomic cosets:
        C_0 ∪ C_1
Generator polynomial:
        x^2 + (α + 1)*x + α
Generator matrix: 5 × 7
        α α + 1 1 0 0 0 0
        0 α α + 1 1 0 0 0
        0 0 α α + 1 1 0 0
        0 0 0 α α + 1 1 0
        0 0 0 0 α α + 1 1

julia> ReedSolomonCode(13, 5, 1)
[12, 8, ≥5; 1]_13 Reed Solomon code.
13-Cyclotomic cosets:
        C_1 ∪ C_2 ∪ C_3 ∪ C_4
Generator polynomial:
        x^4 + 9*x^3 + 7*x^2 + 2*x + 10
Generator matrix: 8 × 12
        10 2 7 9 1 0 0 0 0 0 0 0
        0 10 2 7 9 1 0 0 0 0 0 0
        0 0 10 2 7 9 1 0 0 0 0 0
        0 0 0 10 2 7 9 1 0 0 0 0
        0 0 0 0 10 2 7 9 1 0 0 0
        0 0 0 0 0 10 2 7 9 1 0 0
        0 0 0 0 0 0 10 2 7 9 1 0
        0 0 0 0 0 0 0 10 2 7 9 1

Missing docstring.

Missing docstring for QuadraticResidueCode. Check Documenter's build log for details.

Attributes

CodingTheory.splittingfield — Function

splittingfield(C::AbstractCyclicCode)

Return the splitting field of the generator polynomial.

CodingTheory.polynomialring — Function

polynomialring(C::AbstractCyclicCode)

Return the polynomial ring of the generator polynomial.

CodingTheory.primitiveroot — Function

primitiveroot(C::AbstractCyclicCode)

Return the primitive root of the splitting field.

CodingTheory.offset — Function

offset(C::AbstractBCHCode)

Return the offset of the BCH code.

CodingTheory.designdistance — Function

designdistance(C::AbstractBCHCode)

Return the design distance of the BCH code.

CodingTheory.qcosets — Function

qcosets(C::AbstractCyclicCode)

Return the q-cyclotomic cosets of the cyclic code.

CodingTheory.qcosetsreps — Function

qcosetsreps(C::AbstractCyclicCode)

Return the set of representatives for the q-cyclotomic cosets of the cyclic code.

CodingTheory.definingset — Function

definingset(C::AbstractCyclicCode)

Return the defining set of the cyclic code.

definingset(nums::Vector{Int}, q::Int, n::Int, flat::Bool=true)

Returns the set of q-cyclotomic cosets of the numbers in nums modulo n.

Notes

If flat is set to true, the result will be a single flattened and sorted array.

Base.zeros — Function

zeros(C::AbstractCyclicCode)

Return the zeros of C.

CodingTheory.generatorpolynomial — Function

generatorpolynomial(C::AbstractCyclicCode)

Return the generator polynomial of the cyclic code.

CodingTheory.paritycheckpolynomial — Function

paritycheckpolynomial(C::AbstractCyclicCode)

Return the parity-check polynomial of the cyclic code.

CodingTheory.idempotent — Function

idempotent(C::AbstractCyclicCode)

Return the idempotent (polynomial) of the cyclic code.

CodingTheory.BCHbound — Function

BCHbound(C::AbstractCyclicCode)

Return the BCH bound for C.

CodingTheory.isnarrowsense — Function

isnarrowsense(C::AbstractBCHCode)

Return true if the BCH code is narrowsense.

CodingTheory.isreversible — Function

isreversible(C::AbstractCyclicCode)

Return true if the cyclic code is reversible.

CodingTheory.isdegenerate — Function

isdegenerate(C::AbstractCyclicCode)

Return true if the cyclic code is degenerate.

Notes

A cyclic code is degenerate if the parity-check polynomial divides x^r - 1 for

some r less than the length of the code.

Hecke.isprimitive — Function

isprimitive(C::AbstractBCHCode)

Return true if the BCH code is primitive.

CodingTheory.isantiprimitive — Function

isantiprimitive(C::AbstractBCHCode)

Return true if the BCH code is antiprimitive.

Methods

Missing docstring.

Missing docstring for definingset. Check Documenter's build log for details.

CodingTheory.dualdefiningset — Function

dualdefiningset(defset::Vector{Int}, n::Int)

Return the defining set of the dual code of length n and defining set defset.

Missing docstring.

Missing docstring for iscyclic. Check Documenter's build log for details.

Oscar.complement — Function

complement(C::AbstractCyclicCode)

Return the cyclic code whose cyclotomic cosets are the completement of C's.

Base.:∩ — Function

∩(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the intersection code of C1 and C2.

Base.:+ — Function

+(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the addition code of C1 and C2.