Linear Codes

Constructors

Generic linear codes may be constructed in two ways: via a matrix or via a vector space object. If a vector space is used, the basis of the vector space is used as a generator matrix for the code. If the optional parameter parity is set to true, the input is considered a parity-check matrix instead of a generator matrix. At the moment, no convention is used for the zero code and an error is thrown for such imputs. Zero rows are automatically removed from the input but zero columns are not. See the tutorials for usage examples.

CodingTheory.LinearCode — Type

LinearCode(G::CTMatrixTypes, parity::Bool=false)

Return the linear code constructed with generator matrix G. If the optional paramater parity is set to true, a linear code is built with G as the parity-check matrix.

Attributes

Various getter/accessor functions are provided for accessing attributes about the codes. The user is strongly encouraged to use these functions and never to work with the underlying structs directly, as many functions rely on the information in the structs to be in a specific order and don't check if information has been updated.

CodingTheory.field — Function

field(C::AbstractLinearCode)

Return the base ring of the generator matrix.

field(S::AbstractSubsystemCode)

Return the base ring of the code.

Base.length — Function

length(C::AbstractLinearCode)

Return the length of the code.

length(S::AbstractSubsystemCode)
numqubits(S::AbstractSubsystemCode)

Return the length of the code.

Hecke.dimension — Function

dimension(C::AbstractLinearCode)

Return the dimension of the code.

dimension(S::AbstractSubsystemCode)

Return the dimension of the code.

CodingTheory.cardinality — Function

cardinality(C::AbstractLinearCode)

Return the cardinality of the code.

cardinality(S::AbstractSubsystemCode)

Return the cardinality of the stabilizer group of the code.

CodingTheory.rate — Function

rate(C::AbstractLinearCode)

Return the rate, $R = k/n$, of the code.

rate(S::AbstractSubsystemCode)

Return the rate, R = k/n, of the code.

If the linear code was created by passing in a generator (parity-check) matrix, then this matrix is stored in addition to the standard form. Note that this matrix is potentially over complete (has more rows than its rank). The standard form is returned when the optional parameter standform is set to true. Some code families are not constructed using these matrices. In these cases, the matrices are initially missing and are computed and cached when these functions are called for the first time. Direct access to the underlying structs is not recommended.

CodingTheory.generatormatrix — Function

generatormatrix(C::AbstractLinearCode, standform::Bool=false)

Return the generator matrix of the code.

Notes

If the optional parameter standform is set to true, the standard form of the generator matrix is returned instead.

CodingTheory.paritycheckmatrix — Function

paritycheckmatrix(C::AbstractLinearCode, standform::Bool=false)

Return the parity-check matrix of the code.

Notes

If the optional parameter standform is set to true, the standard form of the parity-check matrix is returned instead.

Recall that putting the matrix into standard form may require column permutations. If this is the case, the column permutation matrix $P$ such that $\mathrm{rowspace}(G) = \mathrm{rowspace}(G_\mathrm{stand} * P)$ may be accessed using the following function. If no column permutations are required, this returns missing.

CodingTheory.standardformpermutation — Function

standardformpermutation(C::AbstractLinearCode)

Return the permutation matrix required to permute the columns of the code matrices to have the same row space as the matrices in standard form. Returns missing is no such permutation is required.

The minimum distance of some code families are known and are set during construction. The minimum distance is automatically computed in the constructor for codes which are deemed "small enough". Otherwise, the minimum distance is missing. Primitive bounds on the minimum distance are given by

CodingTheory.minimumdistancelowerbound — Function

minimumdistancelowerbound(C::AbstractLinearCode)

Return the current lower bound on the minimum distance of C.

CodingTheory.minimumdistanceupperbound — Function

minimumdistanceupperbound(C::AbstractLinearCode)

Return the current upper bound on the minimum distance of C.

If the minimum distance of the code is known, the following functions return useful properties; otherwise they return missing.

CodingTheory.relativedistance — Function

relativedistance(C::AbstractLinearCode)

Return the relative minimum distance, $\delta = d / n$ of the code if $d$ is known; otherwise return missing.

relativedistance(S::AbstractSubsystemCode)

Return the relative minimum distance, δ = d / n of the code if d is known, otherwise errors.

CodingTheory.genus — Function

genus(C::AbstractLinearCode)

Return the genus, $n + 1 - k - d$, of the code.

CodingTheory.isMDS — Function

isMDS(C::AbstractLinearCode)

Return true if code is maximum distance separable (MDS).

CodingTheory.numbercorrectableerrors — Function

numbercorrectableerrors(C::AbstractLinearCode)

Return the number of correctable errors for the code.

Notes

The number of correctable errors is $t = \floor{(d - 1) / 2}$.

The minimum distance and its bounds may be manually set as well. Nothing is done to check this value for correctness.

CodingTheory.setdistancelowerbound! — Function

setdistancelowerbound!(C::AbstractLinearCode, l::Int)

Set the lower bound on the minimum distance of C, if l is better than the current bound.

CodingTheory.setdistanceupperbound! — Function

setdistanceupperbound!(C::AbstractLinearCode, u::Int)

Set the upper bound on the minimum distance of C, if u is better than the current bound.

CodingTheory.setminimumdistance! — Function

setminimumdistance(C::AbstractLinearCode, d::Int)

Set the minimum distance of the code to d.

Notes

The only check done on the value of d is that $1 \leq d \leq n$.

setminimumdistance(S::AbstractSubsystemCode, d::Int)

Set the minimum distance of the code to d.

Notes

The only check done on the value of d is that 1 ≤ d ≤ n.

Methods

CodingTheory.Singletonbound — Function

Singletonbound(n::Int, a::Int)

Return the Singleton bound $d \leq n - k + 1$ or $k \leq n - d + 1$ depending on the interpretation of a.

Singletonbound(C::AbstractLinearCode)

Return the Singleton bound on the minimum distance of the code ($d \leq n - k + 1$).

CodingTheory.encode — Function

encode(C::AbstractLinearCode, v::Union{CTMatrixTypes, Vector{Int}})

Return v * G, where G is the generator matrix of C.

CodingTheory.syndrome — Function

syndrome(C::AbstractLinearCode, v::Union{CTMatrixTypes, Vector{Int}})

Return Hv, where H is the parity-check matrix of C.

Base.in — Function

in(v::Union{CTMatrixTypes, Vector{Int}}, C::AbstractLinearCode)

Return whether or not v is a codeword of C.

Base.:⊆ — Function

⊆(C1::AbstractLinearCode, C2::AbstractLinearCode)
⊂(C1::AbstractLinearCode, C2::AbstractLinearCode)
issubcode(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return whether or not C1 is a subcode of C2.

⊆(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
⊂(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
issubcode(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return whether or not C1 is a subcode of C2.

CodingTheory.areequivalent — Function

areequivalent(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return true if C1 and C2 are permutation equivalent codes.

areequivalent(S1::T, S2::T) where T <: AbstractSubsystemCode

Return true if the codes are equivalent as symplectic vector spaces.

Note

This is not intended to detect if S1 and S2 are permutation equivalent.

Missing docstring.

Missing docstring for dual. Check Documenter's build log for details.

CodingTheory.isselfdual — Function

isselfdual(C::AbstractLinearCode)

Return true if areequivalent(C, dual(C)).

isselfdual(C::AbstractCyclicCode)

Return whether or not C == dual(C).

CodingTheory.isselforthogonal — Function

isselforthogonal(C::AbstractLinearCode)
isweaklyselfdual(C::AbstractLinearCode)
isEuclideanselforthogonal(C::AbstractLinearCode)

Return true if C ⊆ dual(C).

CodingTheory.isdualcontaining — Function

isdualcontaining(C::AbstractLinearCode)
isEuclideandualcontaining(C::AbstractLinearCode)

Return true if dual(C) ⊆ C.

CodingTheory.hull — Function

hull(C::AbstractLinearCode)
Euclideanhull(C::AbstractLinearCode)

Return the (Euclidean) hull of C and its dimension.

Notes

The hull of a code is the intersection of it and its dual.

CodingTheory.isLCD — Function

isLCD(C::AbstractLinearCode)

Return true if C is linear complementary dual.

Notes

A code is linear complementary dual if the dimension of hull(C) is zero.

CodingTheory.Hermitiandual — Function

Hermitiandual(C::AbstractLinearCode)

Return the Hermitian dual of a code defined over a quadratic extension.

Hermitiandual(C::MatrixProductCode)

Return the Hermitian dual of C. If the dual is also a matrix product code, a matrix product code will be returned.

CodingTheory.isHermitianselfdual — Function

isHermitianselfdual(C::AbstractLinearCode)

Return true if areequivalent(C, Hermitiandual(C)).

CodingTheory.isHermitianselforthogonal — Function

isHermitianselforthogonal(C::AbstractLinearCode)
isHermitianweaklyselfdual(C::AbstractLinearCode)

Return true if C ⊆ Hermitiandual(C).

CodingTheory.isHermitiandualcontaining — Function

isHermitiandualcontaining(C::AbstractLinearCode)

Return true if Hermitiandual(C) ⊆ C.

CodingTheory.Hermitianhull — Function

Hermitianhull::AbstractLinearCode)

Return the Hermitian hull of C and its dimension.

Notes

The Hermitian hull of a code is the intersection of it and its Hermitian dual.

CodingTheory.isHermitianLCD — Function

isHermitianLCD(C::AbstractLinearCode)

Return true if C is linear complementary Hermitian dual.

Notes

A code is linear complementary Hermitian dual if the dimension of Hermitianhull(C) is zero.

Base.iseven — Function

iseven(C::AbstractLinearCode)

Return true if C is even.

CodingTheory.isdoublyeven — Function

isdoublyeven(C::AbstractLinearCode)

Return true if C is doubly-even.

CodingTheory.istriplyeven — Function

istriplyeven(C::AbstractLinearCode)

Return true if C is triply-even.

CodingTheory.characteristicpolynomial — Function

characteristicpolynomial(C::AbstractLinearCode)

Return the characteristic polynomial of C.

Notes

The characteristic polynomial is defined in [Lin1999]_

AbstractAlgebra.VectorSpace — Function

VectorSpace(C::AbstractLinearCode)

Return the code C as a vector space object.

Missing docstring.

Missing docstring for words. Check Documenter's build log for details.