LDPC Codes

CodingTheory.LDPCCode — Method

LDPCCode(C::AbstractLinearCode)

Return the LDPC code given by C.

LDPC codes are typically required to have a matrix density of less than 1%.

CodingTheory.LDPCCode — Method

LDPCCode(H::fq_nmod_mat)

Return the LDPC code defined by the parity-check matrix H.

LDPC codes are typically required to have a matrix density of less than 1%.

CodingTheory.Tannergraph — Method

Tannergraph(C::AbstractLDPCCode)

Return the Tanner graph of C as a Figure object.

CodingTheory.Tannergraph — Method

Tannergraph(H::fq_nmod_mat)

Return the Tanner graph of the matrix H as a Figure object.

CodingTheory.checkdegreedistribution — Method

checkdegreedistribution(C::AbstractLDPCCode)

Return the check node degree distribution of C.

CodingTheory.checkdegreepolynomial — Method

checkdegreepolynomial(C::AbstractLDPCCode)

Return the check degree polynomial of C.

CodingTheory.columnbound — Method

columnbound(C::AbstractLDPCCode)

Return the column bound c of the (c, r)-LDPC code C.

CodingTheory.columnrowbounds — Method

columnrowbounds(C::AbstractLDPCCode)

Return the column and row bounds c, r of the (c, r)-LDPC code C.

CodingTheory.degreedistributions — Method

degreedistributions(C::AbstractLDPCCode)

Return the variable and check node degree distributions of C.

CodingTheory.degreedistributionssplot — Method

degreedistributionssplot(C::AbstractLDPCCode)

Return a bar plot of the column and row degree distributions of C.

CodingTheory.density — Method

density(C::AbstractLDPCCode)

Return the density of the parity-check matrix of C.

CodingTheory.isregular — Method

isregular(C::AbstractLDPCCode)

Return true if the C is a regular LDPC code.

An LDPC is regular if all the column degrees and equal and all the row degrees are equal.

CodingTheory.rowbound — Method

rowbound(C::AbstractLDPCCode)

Return the row bound r of the (c, r)-LDPC code C.

CodingTheory.variabledegreedistribution — Method

variabledegreedistribution(C::AbstractLDPCCode)

Return the variable node degree distribution of C.

CodingTheory.variabledegreepolynomial — Method

variabledegreepolynomial(C::AbstractLDPCCode)

Return the variable degree polynomial of C.