LDPC Noise Channels

Constructors

CodingTheory.BinaryErasureChannel — Type

BinaryErasureChannel(ε::Float64)
BEC(ε::Float64)

Return the binary erasure channel with erasure probability ε.

CodingTheory.BinarySymmetricChannel — Type

BinarySymmetricChannel(p::Float64)
BSC(p::Float64)

Return the binary symmetric channel with crossover probability p.

CodingTheory.BAWGNChannel — Type

BAWGNChannel(σ::Float64)
BAWGNC(σ::Float64)

Return the binary (input) additive white Gaussian noise channel with standard deivation σ (noise variance σ^2).

Attributes

CodingTheory.erasure_probability — Function

erasure_probability(Ch::BinaryErasureChannel)

Return the erasure probability of the binary erasure channel.

CodingTheory.crossover_probability — Function

crossover_probability(Ch::BinarySymmetricChannel)

Return the crossover probability of the binary symmetric channel.

CodingTheory.standard_deviation — Function

standard_deviation(Ch::BAWGNChannel)

Return the standard deviation of the BAWGN channel.

CodingTheory.variance — Function

variance(Ch::BAWGNChannel)

Return the variance of the BAWGN channel.

Methods

DataStructures.capacity — Function

capacity(Ch::AbstractClassicalNoiseChannel)

Return the capacity of the noise channel.