charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

The orthogonality of the stabilizers are automatically checked and will error upon failure.

assumed to be in symplectic form over the base field.

This is intended to be a simple function wrapper for typeof(S) since the

constructor for SubsystemCode automatically returns a SubsystemCodeCSS if possible. Manually changing the elements of the struct S without using the helper functions provided here is therefore not recommended.

Stabilizer Codes

CodingTheory.StabilizerCode — Type

StabilizerCode(S::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing)

Return the stabilizer code whose stabilizers is determined by S and signs by charvec.

CodingTheory.StabilizerCode — Method

StabilizerCode(S::AbstractSubsystemCode)

Return the stabilizer code from the subsystem code by promoting all of the gauge operators to logical operators.

CodingTheory.StabilizerCode — Method

StabilizerCode(SPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return the stabilizer code whose stabilizers are determined by the vector of Pauli strings SPauli and signs by charvec.

Notes

Any +/- 1 characters in front of each stabilizer are stripped. No check is done to make sure these signs agree with the ones computed using the character vector.

CodingTheory.StabilizerCodeCSS — Type

StabilizerCodeCSS(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)
CSSCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the CSS code given by the CSS construction on two linear codes C1 and C2 with C2 ⊆ C1 and whose signs by charvec.

Notes

If C1 = [n, k1, d1] and C2 = [n, k2, d2], the resulting code has dimension k = k1 - k2 and minimum distance d >= min(d1, d2^⟂). The X stabilizers are given by the parity-check matrix of C2^⟂, H(C2^⟂), and the Z stabilizers by H(C1).

CodingTheory.StabilizerCodeCSS — Type

StabilizerCodeCSS(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)
CSSCode(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return the CSS code given by the CSS construction on a self-orthogonal linear code C, i.e., C ⊆ C^⟂, and whose signs by charvec.

Notes

Setting C1 = C^⟂ and C2 = C, the resulting code has dimension k = k1 - k2 and minimum distance d >= min(d1, d2^⟂). The X stabilizers are given by the parity-check matrix of C2^⟂, H(C2^⟂), and the Z stabilizers by H(C1).

CodingTheory.StabilizerCodeCSS — Type

StabilizerCodeCSS(S::StabilizerCode)
CSSCode(S::StabilizerCode)

Return the [[2n, 2k, S.d <= d <= 2 S.d]] CSS code derived by splitting the stabilizers of S.

CodingTheory.StabilizerCodeCSS — Method

StabilizerCodeCSS(Xmatrix::fq_nmod_mat, Zmatrix::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing)
CSSCode(Xmatrix::fq_nmod_mat, Zmatrix::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing)

Return a CSS code whose X-stabilizers are given by Xmatrix, Z-stabilizers by Zmatrix, and signs by charvec.

CodingTheory.StabilizerCodeCSS — Method

StabilizerCodeCSS(SPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}
CSSCode(SPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return the CSS code whose stabilizers are determined by the vector of Pauli strings SPauli and signs by charvec.

Notes

Any +/- 1 characters in front of each stabilizer are stripped. No check is done to make sure these signs agree with the ones computed using the character vector.

Subsystem Codes

CodingTheory.SubsystemCode — Type

SubsystemCode(G::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing)

Return the subsystem code whose gauge group is determined by G and signs by charvec.

CodingTheory.SubsystemCode — Type

SubsystemCode(S::fq_nmod_mat, L::CTMatrixTypes, G::CTMatrixTypes, charvec::Union{Vector{nmod}, Missing}=missing)

Return the subsystem code whose stabilizers are given by S, (bare) logical operators by L, gauge operators (not including stabilizers) by G, and signs by charvec.

CodingTheory.SubsystemCode — Method

SubsystemCode(SPauli::Vector{T}, LPauli::Vector{T}, GPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return the subsystem code whose stabilizers are given by the vectors of Pauli strings SPauli, (bare) logical operators by LPauli, gauge operators (not including stabilizers) by GPauli, and signs by charvec.

CodingTheory.SubsystemCode — Method

SubsystemCode(GPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return the subsystem code whose gauge group is determined by the vector of Pauli strings GPauli and signs by charvec.

Notes

Any +/- 1 characters in front of each stabilizer are stripped. No check is done to make sure these signs agree with the ones computed using the character vector.

Base.length — Method

length(S::AbstractSubsystemCode)
numqubits(S::AbstractSubsystemCode)

Return the length of the code.

CodingTheory.Xsigns — Method

Xsigns(S::AbstractSubsystemCode)

Return the signs of the X stabilizers of the CSS code.

CodingTheory.Xstabilizers — Method

Xstabilizers(S::AbstractSubsystemCode)

Return the X-stabilizer matrix of the CSS code.

CodingTheory.Xsyndrome — Method

Xsyndrome(S::AbstractSubsystemCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the X stabilizers of the CSS code.

CodingTheory.Zsigns — Method

Zsigns(S::AbstractSubsystemCode)

Return the signs of the Z stabilizers of the CSS code.

CodingTheory.Zstabilizers — Method

Zstabilizers(S::AbstractSubsystemCode)

Return the Z-stabilizer matrix of the CSS code.

CodingTheory.Zsyndrome — Method

Zsyndrome(S::AbstractSubsystemCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the Z stabilizers of the CSS code.

CodingTheory.allstabilizers — Method

allstabilizers(S::AbstractSubsystemCode)
elements(S::AbstractSubsystemCode)

Return a vector of all the elements of the stabilizer group of S.

CodingTheory.areequivalent — Method

areequivalent(S1::T, S2::T) where T <: AbstractSubsystemCode

Return true if the codes are equivalent as symplectic vector spaces.

Note

This is not intended to detect if S1 and S2 are permutation equivalent.

CodingTheory.augment — Function

augment(S::AbstractSubsystemCode, row::fq_nmod_mat, verbose::Bool=true)

Return the code created by added row to the stabilizers of S.

Notes

The goal of this function is to track how the logical operators update given the new stabilizer. The unaffected logical operators are kept during the update and only those which don't commute with the new stabilizer are recomputed. Use verbose to better

CodingTheory.cardinality — Method

cardinality(S::AbstractSubsystemCode)

Return the cardinality of the stabilizer group of the code.

CodingTheory.charactervector — Method

charactervector(S::AbstractSubsystemCode)

Return the character vector of the code.

CodingTheory.dressed — Method

dressed(S::AbstractSubsystemCode)
dressedoperators(S::AbstractSubsystemCode
dressedlogicals(S::AbstractSubsystemCode)

Return a vector of pairs generators for the dressed operators of S.

Notes

Here, the dressed operators are the logicals and the gauge operators.

CodingTheory.expurgate — Function

expurgate(S::AbstractStabilizerCode, rows::Vector{Int}, verbose::Bool=true)

Return the code created by removing the stabilizers indexed by rows.

Notes

The goal of this function is to track how the logical operators update through this process. Here, the original logical pairs are kept and an appropriate number of new pairs are added.

CodingTheory.field — Method

field(S::AbstractSubsystemCode)

Return the base ring of the code.

CodingTheory.fixgauge — Method

fixgauge(::HasGauges, S::AbstractSubsystemCode, pair::Int, which::Symbol)

Return the code induced by adding either the first of gaugeoperators(S)[pair] to the stabilizers if which = :X or the second if which = :Z.

CodingTheory.gaugegroup — Method

gaugegroup(S::AbstractSubsystemCode)
gaugegroupmatrix(S::AbstractSubsystemCode)
gaugegeneratorsmatrix(S::AbstractSubsystemCode)
gaugegroupgeneratorsmatrix(S::AbstractSubsystemCode)

Return a matrix giving a (maybe overcomplete) basis for the gauge group.

Notes

Here, this is the stabilizers and the gauge operators.

CodingTheory.gauges — Method

gauges(S::AbstractSubsystemCode)
gaugeoperators(S::AbstractSubsystemCode

Return a vector of gauge operator generator pairs for S.

Notes

Here, gauge operators refers to the gauge group minus the stabilizers.

CodingTheory.gaugesmatrix — Method

gaugesmatrix(S::AbstractSubsystemCode)
gaugeoperatorsmatrix(S::AbstractSubsystemCode)

Return the result of gauges(S) as a vertically concatenated matrix.

CodingTheory.isCSS — Method

isCSS(S::AbstractSubsystemCode)

Return true if S is CSS.

CodingTheory.islogical — Method

islogical(S::AbstractSubsystemCode, v::fq_nmod_mat)

Return true if the vector v is a logical operator for S.

CodingTheory.isovercomplete — Method

isovercomplete(S::AbstractSubsystemCode)

Return true if S has an overcomplete set of stabilizers.

CodingTheory.logicals — Method

logicals(S::AbstractSubsystemCode)
logicaloperators(S::AbstractSubsystemCode)
barelogicals(S::AbstractSubsystemCode)
bare(S::AbstractSubsystemCode)

Return a vector of logical operator generator pairs for S.

CodingTheory.logicalsmatrix — Method

logicalsmatrix(S::AbstractSubsystemCode)

Returns the result of logicals(S) as a vertically concatenated matrix.

CodingTheory.logicalsstandardform — Method

logicalsstandardform(S::AbstractSubsystemCode)

Return the a matrix of logical operators as determined by the stabilizers in standard form.

CodingTheory.numXstabs — Method

numXstabs(S::AbstractSubsystemCode)

Return the number of X stabilizers of the CSS code.

CodingTheory.numZstabs — Method

numZstabs(S::AbstractSubsystemCode)

Return the number of Z stabilizers of the CSS code.

CodingTheory.permutecode! — Method

permutecode!(S::AbstractSubsystemCode, σ::Union{PermGroupElem, Perm{Int}, Vector{Int}})

Permute the code S by σ.

Notes

If σ is a vector, it is interpreted as the desired column order for the generator matrix of C.

CodingTheory.permutecode — Method

permutecode(S::AbstractSubsystemCode, σ::Union{PermGroupElem, Perm{Int}, Vector{Int}})

Return the code permuted by σ.

Notes

If σ is a vector, it is interpreted as the desired column order for the generator matrix of C.

CodingTheory.printallstabilizers — Method

printallstabilizers(S::AbstractSubsystemCode)
printallelements(S::AbstractSubsystemCode)

Print all the elements of the stabilizer group of S.

CodingTheory.promotegaugestological! — Method

promotegaugestological(S::AbstractSubsystemCode, pairs::Vector{Int})
promotegaugestological(S::AbstractSubsystemCode, pairs::Vector{Int})

Add the gauge pairs in pairs to the logical operators.

CodingTheory.promotelogicalstogauge! — Method

promotelogicalstogauge(S::AbstractSubsystemCode, pairs::Vector{Int})
promotelogicalstogauge!(S::AbstractSubsystemCode, pairs::Vector{Int})

Add the logical pairs in pairs to the gauge operators.

CodingTheory.rate — Method

rate(S::AbstractSubsystemCode)

Return the rate, R = k/n, of the code.

CodingTheory.relativedistance — Method

relativedistance(S::AbstractSubsystemCode)

Return the relative minimum distance, δ = d / n of the code if d is known, otherwise errors.

CodingTheory.setXstabilizers! — Method

setXstabilizers(S::AbstractSubsystemCode, Xstabs::fq_nmod_mat, trimmed::Bool=true)
setXstabilizers!(S::AbstractSubsystemCode, Xstabs::fq_nmod_mat, trimmed::Bool=true)

Set the X stabilizers of S to Xstabs. If trimmed is true, Xstabs are assumed to have length(S) columns; otherwise, they are assumed to be in symplectic form with 2 * length(S) columns.

Notes

A check is done to make sure stabs are equivalent to the current set of stabilizers.

CodingTheory.setZstabilizers! — Method

setZstabilizers(S::AbstractSubsystemCode, Zstabs::fq_nmod_mat, trimmed::Bool=true)
setZstabilizers!(S::AbstractSubsystemCode, Zstabs::fq_nmod_mat, trimmed::Bool=true)

Set the Z stabilizers of S to Zstabs. If trimmed is true, Zstabs are assumed to have length(S) columns; otherwise, they are assumed to be in symplectic form with 2 * length(S) columns.

Notes

A check is done to make sure stabs are equivalent to the current set of stabilizers.

CodingTheory.setlogicals! — Method

setlogicals(S::AbstractSubsystemCode, L::fq_nmod_mat)
setlogicals!(S::AbstractSubsystemCode, L::fq_nmod_mat)

Set the logical operators of S to L.

Notes

A check is done to make sure L are eqivalent to the current set of logicals (up to stabilizers).

CodingTheory.setminimumdistance! — Method

setminimumdistance(S::AbstractSubsystemCode, d::Int)

Set the minimum distance of the code to d.

Notes

The only check done on the value of d is that 1 ≤ d ≤ n.

CodingTheory.setsigns! — Method

setsigns(S::AbstractSubsystemCode, charvec::Vector{nmod})
setsigns!(S::AbstractSubsystemCode, charvec::Vector{nmod})

Set the character vector of S to charvec and update the signs.

CodingTheory.setstabilizers! — Method

setstabilizers(S::AbstractSubsystemCode, stabs::fq_nmod_mat)
setstabilizers!(S::AbstractSubsystemCode, stabs::fq_nmod_mat)

Set the stabilizers of S to stabs.

Notes

A check is done to make sure stabs are equivalent to the current set of stabilizers.

CodingTheory.signs — Method

signs(S::AbstractSubsystemCode)

Return the signs of the stabilizers of the code.

CodingTheory.stabilizers — Function

stabilizers(S::AbstractSubsystemCode, standform::Bool=false)

Return the stabilizer matrix of the code.

Notes

If the optional parameter standform is set to true, the standard form of the stabilizer matrix is returned instead.

CodingTheory.standardformA — Method

standardformA(S::AbstractSubsystemCode)

Return the named matrix A from the standard form of the stabilizer matrix.

CodingTheory.standardformA1 — Method

standardformA1(S::AbstractSubsystemCode)

Return the named matrix A1 from the standard form of the stabilizer matrix.

CodingTheory.standardformA2 — Method

standardformA2(S::AbstractSubsystemCode)

Return the named matrix A2 from the standard form of the stabilizer matrix.

CodingTheory.standardformB — Method

standardformB(S::AbstractSubsystemCode)

Return the named matrix B from the standard form of the stabilizer matrix.

CodingTheory.standardformC1 — Method

standardformC1(S::AbstractSubsystemCode)

Return the named matrix C1 from the standard form of the stabilizer matrix.

CodingTheory.standardformC2 — Method

standardformC2(S::AbstractSubsystemCode)

Return the named matrix C2 from the standard form of the stabilizer matrix.

CodingTheory.standardformD — Method

standardformD(S::AbstractSubsystemCode)

Return the named matrix D from the standard form of the stabilizer matrix.

CodingTheory.standardformE — Method

standardformE(S::AbstractSubsystemCode)

Return the named matrix E from the standard form of the stabilizer matrix.

CodingTheory.standardformpermutation — Method

standardformpermutation(S::AbstractSubsystemCode)

Return the permutation matrix required to permute the columns of the code matrices to have the same row space as the matrices in standard form. Returns missing is no such permutation is required.

CodingTheory.swapXZgaugeoperators! — Method

swapXZgaugeoperators!(S::AbstractSubsystemCode, pairs::Vector{Int})

Swap the X and Z gauge operators specified by pairs.

CodingTheory.swapXZlogicals! — Method

swapXZlogicals!(S::AbstractSubsystemCode, pairs::Vector{Int})

Swap the X and Z logicals specified by pairs.

CodingTheory.syndrome — Method

syndrome(S::AbstractSubsystemCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the stabilizers of S.

Hecke.dimension — Method

dimension(S::AbstractSubsystemCode)

Return the dimension of the code.

Graph States

CodingTheory.ClusterState — Method

ClusterState(w::Int, h::Int)

Return the cluster state (graph state) on the rectangular lattice with width w and height h.

CodingTheory.GraphState — Method

GraphState(G::SimpleGraph{Int64})

Return the graph state defined by the graph G.