Weight Reduction

See the weight reduction tutorial for a more detailed explanation of this function.

CodingTheory.copying — Function

copying(H_X::CTMatrixTypes, H_Z::CTMatrixTypes; method::Symbol = :Hastings, target_q_X::Int = 3)

Return the result of copying on H_X and H_Z using either the Hastings, reduced, or targeted methods.

copying(S::AbstractStabilizerCode, method::Symbol = :Hastings, target_q_X::Int = 3)

Return the result of copying on S using either the Hastings, reduced, or targeted methods.

CodingTheory.copying_as_coning — Function

copying_as_coning(H_X::CTMatrixTypes, H_Z::CTMatrixTypes; method::Symbol = :Hastings, target_q_X::Int = 3)

Return the result of copying on H_X and H_Z using either the Hastings, reduced, or targeted methods by using the mapping cone.

copying_as_coning(S::AbstractStabilizerCode, method::Symbol = :Hastings, target_q_X::Int = 3)

Return the result of copying on S using either the Hastings, reduced, or targeted methods by using the mapping cone.

CodingTheory.gauging — Function

gauging(H_X::CTMatrixTypes, H_Z::CTMatrixTypes)

Return the result of gauging on H_X and H_Z.

gauging(S::AbstractStabilizerCode)

Return the result of gauging on S.

CodingTheory.gauging_as_coning — Function

gauging_as_coning(H_X::CTMatrixTypes, H_Z::CTMatrixTypes)

Return the result of gauging on H_X and H_Z by using the mapping cone.

gauging_as_coning(S::AbstractStabilizerCode)

Return the result of gauging on S by using the mapping cone.

CodingTheory.thickening_and_choose_heights — Function

thickening_and_choose_heights(H_X::CTMatrixTypes, H_Z::CTMatrixTypes, l::Integer, heights::Vector{Int})

Return the result of thickening and choosing heights on H_X and H_Z.

thickening_and_choose_heights(S::AbstractStabilizerCode, l::Integer, heights::Vector{Int})

Return the result of thickening and choosing heights on S.

CodingTheory.coning — Function

coning(H_X::T, H_Z::T, row_indices::AbstractVector{Int}; l::Int = 0, target_q_X::Int = 3) where T <: CTMatrixTypes

Return the result of coning on H_X and H_Z by reducing the Z stabilizers in row_indices and using the optional arguments l and target_q_X for an optional round of thickening and choosing heights.

coning(S::AbstractStabilizerCode, row_indices::AbstractVector{Int}; l::Int = 0, target_q_X::Int = 3) where T <: CTMatrixTypes

Return the result of coning on S by reducing the Z stabilizers in row_indices and using the optional arguments l and target_q_X for an optional round of thickening and choosing heights.

CodingTheory.quantum_weight_reduction — Function

weight_reduction(S::AbstractStabilizerCode, copying_type::Symbol=:Hastings, copying_target::Int = 3, l1::Int, heights::Vector{Int}, l2::Int = 1, target_q_X::Int = 3, seed::Union{Nothing, Int} = nothing)
quantum_weight_reduction(S::AbstractStabilizerCode, copying_type::Symbol=:Hastings, copying_target::Int = 3, l1::Int, heights::Vector{Int}, l2::Int = 1, target_q_X::Int = 3, seed::Union{Nothing, Int} = nothing)

Return the weight-reduced CSS code of S.