Reed-Muller Codes

CodingTheory.ReedMullerCode — Method

ReedMullerCode(q::Integer, r::Integer, m::Integer)

Return the RM(r, m) Reed-Muller code over GF(q).

CodingTheory.ReedMullergeneratormatrix — Method

ReedMullergeneratormatrix(q::Integer, r::Integer, m::Integer)

Return the recursive form of the generator matrix for the RM(r, m) Reed-Muller code over GF(q).

CodingTheory.dual — Method

dual(C::ReedMullerCode)

Return the dual of the Reed-Muller code C.

CodingTheory.entrywiseproductcode — Method

entrywiseproductcode(C::ReedMullerCode, D::ReedMullerCode)
*(C::ReedMullerCode, D::ReedMullerCode)
Schurproductcode(C::ReedMullerCode, D::ReedMullerCode)
Hadamardproductcode(C::ReedMullerCode, D::ReedMullerCode)
componentwiseproductcode(C::ReedMullerCode, D::ReedMullerCode)

Return the entrywise product of C and D.

Note that this is known to often be the full ambient space.

GroupsCore.order — Method

order(C::ReedMullerCode)
RMr(C::ReedMullerCode)

Return the order, r, of the RM(r, m) Reed-Muller code.