Cyclotomic Cosets

CodingTheory.CyclotomicMod.allcyclotomiccosets — Function

allcyclotomiccosets(q::Integer, n::Integer, tosort::Bool=true, verbose::Bool=false)

Return all q-cyclotomic cosets modulo n.

If the optional parameter tosort is set to false, the result will not be sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.CyclotomicMod.complementqcosets — Method

complementqcosets(q::Integer, n::Integer, qcosets::Vector{Vector{Int64}})

Return the complement of the q-cyclotomic cosets modulo n of qcosets.

CodingTheory.CyclotomicMod.cyclotomiccoset — Function

cyclotomiccoset(x::Integer, q::Integer, n::Integer, tosort::Bool=true, verbose::Bool=false)

Return the q-cyclotomic coset of x modulo n.

If the optional parameter tosort is set to false, the result will not be sorted. If the optional parameter verbose is set to true, the result will pretty print.

CodingTheory.CyclotomicMod.dualqcosets — Method

dualqcosets(q::Integer, n::Integer, qcosets::Vector{Vector{Int64}})

Return the dual of the q-cyclotomic cosets modulo n of qcosets.

CodingTheory.CyclotomicMod.ord — Method

ord(n::Integer, q::Integer)

Return the order of n mod q.

CodingTheory.CyclotomicMod.qcosetpairings — Method

qcosetpairings(arr::Vector{Vector{Int64}}, n::Integer)

Return the .

Cyclic Codes

CodingTheory.CyclicCodeMod.BCHCode — Type

BCHCode(q::Integer, n::Integer, δ::Integer, b::Integer=0, verify::Bool=true)

Return the BCHCode of length n over GF(q) with design distance δ and offset b.

This function will auto determine if the constructed code is Reed-Solomon and call the appropriate constructor. If the optional parameter verify is set to true, basic checks, such as checking g(x)h(x) == x^n - 1` and checking for column swaps in the standard form, are done to ensure correctness.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> B = BCHCode(q, n, δ, b)
[15, 5, ≥7; 1]_2 BCH code over splitting field GF(16).
2-Cyclotomic cosets:
        C_1 ∪ C_3 ∪ C_5
Generator polynomial:
        x^10 + x^8 + x^5 + x^4 + x^2 + x + 1
Generator matrix: 5 × 15
        1 1 1 0 1 1 0 0 1 0 1 0 0 0 0
        0 1 1 1 0 1 1 0 0 1 0 1 0 0 0

        0 0 1 1 1 0 1 1 0 0 1 0 1 0 0
        0 0 0 1 1 1 0 1 1 0 0 1 0 1 0
        0 0 0 0 1 1 1 0 1 1 0 0 1 0 1

CodingTheory.CyclicCodeMod.CyclicCode — Type

CyclicCode(q::Integer, n::Integer, cosets::Vector{Vector{Int64}}, verify::Bool=true)

Return the CyclicCode of length n over GF(q) with q-cyclotomic cosets cosets.

This function will auto determine if the constructed code is BCH or Reed-Solomon and call the appropriate constructor. If the optional parameter verify is set to true, basic checks, such as checking g(x)h(x) == x^n - 1` and checking for column swaps in the standard form, are done to ensure correctness.

Examples

julia> q = 2; n = 15; b = 3; δ = 4;
julia> cosets = definingset([i for i = b:(b + δ - 2)], q, n, false);
julia> C = CyclicCode(q, n, cosets)

CodingTheory.CyclicCodeMod.ReedSolomonCode — Type

ReedSolomonCode(q::Integer, δ::Integer, b::Integer=0, verify::Bool=true)

Return the ReedSolomonCode over GF(q) with distance d and offset b.

If the optional parameter verify is set to true, basic checks, such as checking g(x)h(x) == x^n - 1` and checking for column swaps in the standard form, are done to ensure correctness.

Examples

julia> ReedSolomonCode(8, 3, 0)
[7, 5, ≥3; 0]_8 Reed Solomon code.
8-Cyclotomic cosets:
        C_0 ∪ C_1
Generator polynomial:
        x^2 + (α + 1)*x + α
Generator matrix: 5 × 7
        α α + 1 1 0 0 0 0
        0 α α + 1 1 0 0 0
        0 0 α α + 1 1 0 0
        0 0 0 α α + 1 1 0
        0 0 0 0 α α + 1 1

julia> ReedSolomonCode(13, 5, 1)
[12, 8, ≥5; 1]_13 Reed Solomon code.
13-Cyclotomic cosets:
        C_1 ∪ C_2 ∪ C_3 ∪ C_4
Generator polynomial:
        x^4 + 9*x^3 + 7*x^2 + 2*x + 10
Generator matrix: 8 × 12
        10 2 7 9 1 0 0 0 0 0 0 0
        0 10 2 7 9 1 0 0 0 0 0 0
        0 0 10 2 7 9 1 0 0 0 0 0
        0 0 0 10 2 7 9 1 0 0 0 0
        0 0 0 0 10 2 7 9 1 0 0 0
        0 0 0 0 0 10 2 7 9 1 0 0
        0 0 0 0 0 0 10 2 7 9 1 0
        0 0 0 0 0 0 0 10 2 7 9 1

Base.:+ — Method

+(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the addition code of C1 and C2.

Base.:== — Method

==(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return whether or not C1 and C2 have the same fields, lengths, and defining sets.

Base.:∩ — Method

∩(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return the intersection code of C1 and C2.

Base.:⊆ — Method

⊆(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
⊂(C1::AbstractCyclicCode, C2::AbstractCyclicCode)
issubcode(C1::AbstractCyclicCode, C2::AbstractCyclicCode)

Return whether or not C1 is a subcode of C2.

CodingTheory.CyclicCodeMod.complement — Function

complement(C::AbstractCyclicCode, verify::Bool=true)

Return the cyclic code whose cyclotomic cosets are the completement of C's.

If the optional parameter verify is set to true, basic checks, are done to ensure correctness.

CodingTheory.CyclicCodeMod.definingset — Function

definingset(nums::Vector{Int64}, q::Integer, n::Integer, flat::Bool=true)

Returns the set of q-cyclotomic cosets of the numbers in nums modulo n.

If flat is set to true, the result will be a single flattened and sorted array.

CodingTheory.CyclicCodeMod.definingset — Method

definingset(C::AbstractCyclicCode)

Return the defining set of the cyclic code.

CodingTheory.CyclicCodeMod.designdistance — Method

designdistance(C::AbstractCyclicCode)

Return the design distance of the cyclic code.

CodingTheory.CyclicCodeMod.dual — Method

dual(C::AbstractCyclicCode)

Return the dual of the cyclic code C.

Unlike with LinearCode, everything is recomputed here so the proper polynomials and cyclotomic cosets are stored.

CodingTheory.CyclicCodeMod.dualdefiningset — Method

dualdefiningset(defset::Vector{Int64}, n::Integer)

Return the defining set of the dual code of length n and defining set defset.

CodingTheory.CyclicCodeMod.field — Method

field(C::AbstractCyclicCode)

Return the base field of the generator matrix as a Nemo object.

CodingTheory.CyclicCodeMod.finddelta — Method

finddelta(n::Integer, cosets::Vector{Vector{Int64}})

Return the number of consecutive elements of cosets, the offset for this, and a lower bound on the distance of the code defined with length n and cyclotomic cosets cosets.

The lower bound is determined by applying the Hartmann-Tzeng Bound refinement to the BCH bound.

CodingTheory.CyclicCodeMod.generatorpolynomial — Method

generatorpolynomial(C::AbstractCyclicCode)

Return the generator polynomial of the cyclic code as a Nemo object.

CodingTheory.CyclicCodeMod.idempotent — Method

idempotent(C::AbstractCyclicCode)

Return the idempotent (polynomial) of the cyclic code as a Nemo object.

CodingTheory.CyclicCodeMod.isnarrowsense — Method

isnarrowsense(C::AbstractCyclicCode)

Return true if the cyclic code is narrowsense.

CodingTheory.CyclicCodeMod.isprimitive — Method

isprimitive(C::AbstractCyclicCode)

Return true if the cyclic code is primitive.

CodingTheory.CyclicCodeMod.isreversible — Method

isreversible(C::AbstractCyclicCode)

Return true if the cyclic code is reversible.

CodingTheory.CyclicCodeMod.isselfdual — Method

isselfdual(C::AbstractCyclicCode)

Return whether or not C == dual(C).

CodingTheory.CyclicCodeMod.mindistlowerbound — Method

mindistlowerbound(C::AbstractCyclicCode)

Return a lower bound on the minimum distance of the code.

At the moment, this is only the BCH bound with the Hartmann-Tzeng Bound refinement. The minimum distance is returned if known.

CodingTheory.CyclicCodeMod.offset — Method

offset(C::AbstractCyclicCode)

Return the offset of the cyclic code.

CodingTheory.CyclicCodeMod.paritycheckpolynomial — Method

paritycheckpolynomial(C::AbstractCyclicCode)

Return the parity-check polynomial of the cyclic code as a Nemo object.

CodingTheory.CyclicCodeMod.polynomialring — Method

polynomialring(C::AbstractCyclicCode)

Return the polynomial ring of the generator polynomial as a Nemo object.

CodingTheory.CyclicCodeMod.primitiveroot — Method

primitiveroot(C::AbstractCyclicCode)

Return the primitive root of the splitting field as a Nemo object.

CodingTheory.CyclicCodeMod.qcosets — Method

qcosets(C::AbstractCyclicCode)

Return the q-cyclotomic cosets of the cyclic code.

CodingTheory.CyclicCodeMod.qcosetsreps — Method

qcosetsreps(C::AbstractCyclicCode)

Return the set of representatives for the q-cyclotomic cosets of the cyclic code.

CodingTheory.CyclicCodeMod.splittingfield — Method

splittingfield(C::AbstractCyclicCode)

Return the splitting field of the generator polynomial as a Nemo object.