Linear Codes

CodingTheory.LinearCode — Type

LinearCode(G::fq_nmod_mat, parity::Bool=false)

Return the linear code constructed with generator matrix G.

If G is not full rank, a row-reduced form is computed for the generator matrix. The dimension of the code is the number of rows of the full-rank matrix, and the length the number of columns. If the optional paramater parity is set to true, a LinearCode is built with G as the parity check matrix.

Arguments

G: a matrix over a finite field of type FqNmodFiniteField

parity: a boolean

Notes

At the moment, no convention is used for G = 0 and an error is thrown.

Zero columns are not removed.

Row reduction is based on Nemo's rref function. It should be monitored to make sure that this does not introduce column swapping in a future version.

Examples

julia> F, _ = FiniteField(2, 1, "α");
julia> G = matrix(F, 4, 7, [1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 1, 1, 1, 1]);
julia> C = LinearCode(G)
[7, 4]_2 linear code.
Generator matrix: 4 × 7
        1 0 0 0 0 1 1
        0 1 0 0 1 0 1
        0 0 1 0 1 1 0
        0 0 0 1 1 1 1

AbstractAlgebra.VectorSpace — Method

VectorSpace(C::AbstractLinearCode)

Return the code C as a vector space object.

Base.:⊆ — Method

⊆(C1::AbstractLinearCode, C2::AbstractLinearCode)
⊂(C1::AbstractLinearCode, C2::AbstractLinearCode)
issubcode(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return whether or not C1 is a subcode of C2.

A code C1 is a subcode of another code C2 if each row of the generator matrix of C1 is a valid codeword of C2.

Base.in — Method

in(v::Union{fq_nmod_mat, Vector{Int64}}, C::AbstractLinearCode)

Return whether or not v is a codeword of C.

The vector v is a valid codeword of C if and only if the syndrome of v is zero.

Base.length — Method

length(C::AbstractLinearCode)

Return the length of the code.

CodingTheory.:⊕ — Method

⊕(C1::AbstractLinearCode, C2::AbstractLinearCode)
directsum(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊕ C2

Return the direct sum code of C1 and C2.

The direct sum code has generator matrix G1 ⊕ G2 and parity check matrix H1 ⊕ H2.

CodingTheory.:⊗ — Method

⊗(C1::AbstractLinearCode, C2::AbstractLinearCode)
kron(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
tensorproduct(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
directproduct(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2
productcode(C1::AbstractLinearCode, C2::AbstractLinearCode) = C1 ⊗ C2

Return the (direct/tensor) product code of C1 and C2.

The product code has generator matrix G1 ⊗ G2 and parity check matrix H1 ⊗ H2.

Notes

The resulting product is not checked for any zero columns.

CodingTheory.Hermitiandual — Method

Hermitiandual(C::AbstractLinearCode)

Return the Hermitian dual of a code defined over a quadratic extension.

CodingTheory.augment — Method

augment(C::AbstractLinearCode, M::fq_nmod_mat)

Return the code of C whose generator matrix is augmented with M.

Vertically joins the matrix M to the bottom of the generator matrix of C.

CodingTheory.cardinality — Method

cardinality(C::AbstractLinearCode)

Return the cardinality of the code.

No size checking is done on the parameters of the code, returns a BitInt by default.

CodingTheory.codecomplement — Method

codecomplement(C1::AbstractLinearCode, C2::AbstractLinearCode)
quo(C1::AbstractLinearCode, C2::AbstractLinearCode)
quotient(C1::AbstractLinearCode, C2::AbstractLinearCode)
/(C2::AbstractLinearCode, C1::AbstractLinearCode)

Return the code C2 / C1 given C1 ⊆ C2.

Credit to Tommy Hofmann of the AbstractAlgebra/Nemo/Hecke packages for help with debugging and providing the most elegant implementation now used here.

CodingTheory.constructionX — Method

constructionX(C1::AbstractLinearCode, C2::AbstractLinearCode, C3::AbstractLinearCode)

Return the code generated by the construction X procedure.

Let C1 be an [n, k, d], C2 be an [n, k - l, d + e], and C3 be an [m, l, e] linear code with C2 ⊂ C1 be proper. Construction X creates a [n + m, k, d + e] linear code.

CodingTheory.constructionX3 — Method

constructionX3(C1::AbstractLinearCode, C2::AbstractLinearCode, C3::AbstractLinearCode,
    C4::AbstractLinearCode, C5::AbstractLinearCode))

Return the code generated by the construction X3 procedure.

Let C1 = [n, k1, d1], C2 = [n, k2, d2], C3 = [n, k3, d3], C4 = [n4, k2 - k1, d4], and C5 = [n5, k3 - k2, d5] with C1 ⊂ C2 ⊂ C3. Construction X3 creates an [n + n4 + n5, k3, d] linear code with d ≥ min{d1, d2 + d4, d3 + d5}.

CodingTheory.dimension — Method

dimension(C::AbstractLinearCode)

Return the dimension of the code.

CodingTheory.encode — Method

encode(v::Union{fq_nmod_mat, Vector{Int64}}, C::AbstractLinearCode)

Return v * G, where G is the generator matrix of C.

Arguments

v: Either a 1 × k or a k × 1 vector.

CodingTheory.entrywiseproductcode — Method

entrywiseproductcode(C::AbstractLinearCode, D::AbstractLinearCode)
*(C::AbstractLinearCode, D::AbstractLinearCode)
Schurproductcode(C::AbstractLinearCode, D::AbstractLinearCode)
Hadamardproductcode(C::AbstractLinearCode, D::AbstractLinearCode)
componentwiseproductcode(C::AbstractLinearCode, D::AbstractLinearCode)

Return the entrywise product of C and D.

Note that this is known to often be the full ambient space.

CodingTheory.expandedcode — Method

expandedcode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the expanded code of C constructed by exapnding the generator matrix to the subfield K using the provided dual basis for the field of C over K.

No check is done to ensure that basis is indeed a basis for the extension.

CodingTheory.expurgate — Method

expurgate(C::AbstractLinearCode, rows::Vector{Int64})

Return the code of C expuragated at the rows in rows.

Deletes the rows from the generator matrix and then removes any potentially resulting zero columns.

CodingTheory.extend — Method

extend(C::AbstractLinearCode)

Return the extended code of C.

This implementation chooses the most common form of extending a code, which is to add an extra column to the generator matrix such that the sum of the coordinates of each row is 0.

CodingTheory.field — Method

field(C::AbstractLinearCode)

Return the base ring of the generator matrix.

CodingTheory.generatormatrix — Function

generatormatrix(C::AbstractLinearCode, standform::Bool=false)

Return the generator matrix of the code.

If the optional parameter standform is set to true, the standard form of the generator matrix is returned instead.

CodingTheory.genus — Method

genus(C::AbstractLinearCode)

Return the genus, n + 1 - k - d, of the code.

CodingTheory.isMDS — Method

isMDS(C::AbstractLinearCode)

Return true if code is maximum distance separable (MDS).

A linear code is MDS if it saturates the Singleton bound, d = n - k + 1.

CodingTheory.isequivalent — Method

isequivalent(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return true if C1 ⊆ C2 and C2 ⊆ C1.

CodingTheory.isselfdual — Method

isselfdual(C::AbstractLinearCode)

Return true if isequivalent(C, dual(C)).

CodingTheory.isselforthogonal — Method

isselforthogonal(C::AbstractLinearCode)
isweaklyselfdual(C::AbstractLinearCode)

Return true if C ⊆ dual(C).

CodingTheory.juxtaposition — Method

juxtaposition(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the code generated by the horizontal concatenation of the generator matrices of C1 then C2.

CodingTheory.lengthen — Method

lengthen(C::AbstractLinearCode)

Return the lengthened code of C.

This augments the all 1's row and then extends.

CodingTheory.numbercorrectableerrors — Method

numbercorrectableerrors(C::AbstractLinearCode)

Return the number of correctable errors for the code.

The number of correctable errors is t = floor((d - 1) / 2).

CodingTheory.originalgeneratormatrix — Method

originalgeneratormatrix(C::AbstractLinearCode)

Return the original generator matrix used to create the code.

This should not be interpreted as any kind of generator to the code. For example, if G is not a full-rank matrix, then this function acting on C = LinearCode(G) will return G instead of the generator matrix for C. This may be used to diagnose or understand the resulting code after its creation. An important use of this function is in connection with standard code modification procedures such as extending, puncturing, shortening, etc where this returns the matrix on which these procedures have been applied.

CodingTheory.originalparitycheckmatrix — Method

originalparitycheckmatrix(C::AbstractLinearCode)

Return the original parity check matrix used to create the code.

See originalgeneratormatrix.

CodingTheory.paritycheckmatrix — Function

paritycheckmatrix(C::AbstractLinearCode, standform::Bool=false)

Return the parity check matrix of the code.

If the optional parameter standform is set to true, the standard form of the parity check matrix is returned instead.

CodingTheory.permutecode — Method

permutecode(C::AbstractLinearCode, σ::Union{Perm{T}, Vector{T}}) where T <: Integer

Return the code whose generator matrix is C's with the columns permuted by σ.

If σ is a vector, it is interpreted as the desired column order for the generator matrix of C.

CodingTheory.puncture — Method

puncture(C::AbstractLinearCode, cols::Vector{Int64})

Return the code of C punctured at the columns in cols.

Deletes the columns from the generator matrix and then removes any potentially resulting zero rows.

CodingTheory.rate — Method

rate(C::AbstractLinearCode)

Return the rate, `R = k/n', of the code.

CodingTheory.relativedistance — Method

relativedistance(C::AbstractLinearCode)

Return the relative minimum distance, δ = d / n of the code if d is known, otherwise errors.

CodingTheory.setminimumdistance! — Method

setminimumdistance(C::AbstractLinearCode, d::Integer)

Set the minimum distance of the code to d.

The only check done on the value of d is that 1 ≤ d ≤ n.

CodingTheory.shorten — Method

shorten(C::AbstractLinearCode, L::Vector{Int64})

Return the code of C shortened on the indices L.

Shortening is expurgating followed by puncturing. This implementation uses the theorem that the dual of code shortened on L is equal to the puncture of the dual code on L, i.e., dual(puncture(dual(C), L)).

CodingTheory.subcode — Method

subcode(C::AbstractLinearCode, k::Integer)

Return a k-dimensional subcode of C.

CodingTheory.subcode — Method

subcode(C::AbstractLinearCode, rows::Vector{Int64})

Return a subcode of C using the rows of the generator matrix of C listed in rows.

CodingTheory.subfieldsubcode — Method

subfieldsubcode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the subfield subcode code of C over K using the provided dual basis for the field of C over K.

No check is done to ensure that basis is indeed a basis for the extension.

CodingTheory.syndrome — Method

syndrome(v::Union{fq_nmod_mat, Vector{Int64}}, C::AbstractLinearCode)

Return H * v, where H is the parity check matrix of C.

Arguments

v: Either a 1 × k or a k × 1 vector.

CodingTheory.tracecode — Method

tracecode(C::AbstractLinearCode, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the trace code of C over K using the provided dual basis for the field of C over K using Delsarte's theorem.

No check is done to ensure that basis is indeed a basis for the extension.

CodingTheory.upluswvpluswuplusvplusw — Method

upluswvpluswuplusvplusw(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the code generated by the (u + w | v + w | u + v + w)-construction.

Let C1 = [n, k1, d1] and C2 = [n, k2, d2]. This construction produces an [3n, 2k1 + k2] linear code. For binary codes, wt(u + w | v + w | u + v + w) = 2 wt(u ⊻ v) - wt(w) + 4s, where s = |{i | ui = vi = 0, w_i = 1}|.

CodingTheory.uuplusv — Method

uuplusv(C1::AbstractLinearCode, C2::AbstractLinearCode)
Plotkinconstruction(C1::AbstractLinearCode, C2::AbstractLinearCode)

Return the Plotkin- or so-called (u | u + v)-construction with u ∈ C1 and v ∈ C2.

CodingTheory.words — Function

words(C::AbstractLinearCode, onlyprint::Bool=false)
codewords(C::AbstractLinearCode, onlyprint::Bool=false)
elements(C::AbstractLinearCode, onlyprint::Bool=false)

Return the elements of C.

If onlyprint is true, the elements are only printed to the console and not returned.