Miscellaneous Known Stabilizer Codes

CodingTheory.Q1573 — Method

Q1573()

Return the [[15, 7, 3]] quantum Hamming code.

CodingTheory.Q15RM — Method

Q15RM()
Q1513()

Return the [[15, 1, 3]] quantum Reed-Muller code with stabilizers in standard ordering.

CodingTheory.Shorcode — Method

Shorcode()
Q913()

Return the [[9, 1, 3]] Shor code.

CodingTheory.Steanecode — Method

Steanecode()
Q713()

Return the [[7, 1, 3]] Steane code with stabilizers in standard ordering.

CodingTheory.XZZXsurfacecode — Method

XZZXsurfacecode(d::Int)

Return the [[d^2, 1, d]] XZZX surface code.

CodingTheory.fivequbitcode — Method

fivequbitcode()
Q513()

Return the [[5, 1, 3]] perfect qubit stabilizer code.

CodingTheory.rotatedsurfacecode — Method

rotatedsurfacecode(d::Int)

Return the [[d^2, 1, d]] rotated surface code.

CodingTheory.tricolorcode488 — Method

tricolorcode488(d::Int)

Return the 4.8.8 triangular color code of distance d with trellis numbering.

CodingTheory.tricolorcode666 — Method

tricolorcode666(d::Int)

Return the 6.6.6 triangular color code of distance d with trellis numbering.