Stabilizer Codes

CodingTheory.CSSCode — Type

CSSCode(C1::AbstractLinearCode, C2::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return a CSS code using the CSS construction on two linear codes C1 = [n, k1, d1] and C2 = [n, k2, d2] with C2 ⊆ C1.

The resulting code has dimension k = k1 - k2 and minimum distance d >= min(d1, d2^⟂). The X stabilizers are given by the parity-check matrix of C2^⟂, H(C2^⟂), and the Z stabilizers by H(C1).

Arguments

C1: a linear code

C2: a subcode of C1

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.CSSCode — Type

CSSCode(C::AbstractLinearCode, charvec::Union{Vector{nmod}, Missing}=missing)

Return a CSS code using the CSS construction on a self-orthogonal linear code C, i.e., C ⊆ C^⟂.

Setting C1 = C^⟂ and C2 = C, the resulting code has dimension k = k1 - k2 and minimum distance d >= min(d1, d2^⟂). The X stabilizers are given by the parity-check matrix of C2^⟂, H(C2^⟂), and the Z stabilizers by H(C1).

It is often desirable in quantum error correction to work with a set of overcomplete stabilizers. Therefore this constructor does not simplify any provided set of stabilizers. The dimension of the code is computed based on the rank, and the user should not use the matrix dimension of the stabilizers to determine such quantities. Use isovercomplete to determine if an AbstractStabilizerCode is overcomplete.

Arguments

C: a self-orthogonal linear code

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

CodingTheory.CSSCode — Type

CSSCode(Xmatrix::fq_nmod_mat, Zmatrix::fq_nmod_mat, charvec::Union{Vector{nmod}, Missing}=missing)

Return a CSS code using the matrix Xmatrix as the X stabilizer matrix and Zmatrix as the Z stabilizer matrix.

It is often desirable in quantum error correction to work with a set of overcomplete stabilizers. Therefore this constructor does not simplify any provided set of stabilizers. The dimension of the code is computed based on the rank, and the user should not use the matrix dimension of the stabilizers to determine such quantities. Use isovercomplete to determine if an AbstractStabilizerCode is overcomplete.

Arguments

Xmatrix: a matrix over a finite field of type FqNmodFiniteField

Zmatrix: a matrix over a finite field of type FqNmodFiniteField

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

The orthogonality of the X and Z stabilizers are automatically checked and will error upon failure.

CodingTheory.CSSCode — Method

CSSCode(SPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return a CSS code using the vector of Pauli strings SPauli as stabilizers.

It is often desirable in quantum error correction to work with a set of overcomplete stabilizers. Therefore this constructor does not simplify any provided set of stabilizers. The dimension of the code is computed based on the rank, and the user should not use the matrix dimension of the stabilizers to determine such quantities. Use isovercomplete to determine if an AbstractStabilizerCode is overcomplete.

Arguments

SPauli: a vector of Strings or Char vectors containing the letters {I, X, Y, Z}

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

The orthogonality of the X and Z stabilizers are automatically checked and will error upon failure.

Any +/- 1 characters in front of each stabilizer are stripped. No check is done to make sure these signs agree with the ones computed using the character vector.

Will error when the provided strings are not CSS.

CodingTheory.QuantumCode — Method

QuantumCode(Sq2::fq_nmod_mat, symp::Bool=false, charvec::Union{Vector{nmod}, Missing}=missing)

Return a stabilizer code using the matrix Sq2 as the stabilizer matrix.

The matrix Sq2 is assumed to be an n column matrix over the quadratic extension. If the optional parameter symp is set to true, Sq2 is assumed to be in symplectic form over the base field.

It is often desirable in quantum error correction to work with a set of overcomplete stabilizers. Therefore this constructor does not simplify any provided set of stabilizers. The dimension of the code is computed based on the rank, and the user should not use the matrix dimension of the stabilizers to determine such quantities. Use isovercomplete to determine if an AbstractStabilizerCode is overcomplete.

Arguments

Sq2: a matrix over a finite field of type FqNmodFiniteField

symp: a boolean

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

The orthogonality of the stabilizers are automatically checked and will error upon failure.

CodingTheory.QuantumCode — Method

QuantumCode(SPauli::Vector{T}, charvec::Union{Vector{nmod}, Missing}=missing) where T <: Union{String, Vector{Char}}

Return a stabilizer code using the vector of Pauli strings SPauli as stabilizers.

It is often desirable in quantum error correction to work with a set of overcomplete stabilizers. Therefore this constructor does not simplify any provided set of stabilizers. The dimension of the code is computed based on the rank, and the user should not use the matrix dimension of the stabilizers to determine such quantities. Use isovercomplete to determine if an AbstractStabilizerCode is overcomplete.

Arguments

SPauli: a vector of Strings or Char vectors containing the letters {I, X, Y, Z}

charvec: a length 2n vector with elements in the Z/(2p) if chracteristic(field(C1)) is 2 and Z/(p) otherwise. The first n elements specify the exponents of the X phases and second n the exponents of the Z phases; a missing argument will be set to the all-zero vector

Notes

A +1 phase should be entered as 0 since the character vector stores the exponents.

Stabilizer signs are automatically computed given the character vector.

The orthogonality of the stabilizers are automatically checked and will error upon failure.

Any +/- 1 characters in front of each stabilizer are stripped. No check is done to make sure these signs agree with the ones computed using the character vector.

Base.length — Method

length(S::AbstractStabilizerCode)
numqubits(S::AbstractStabilizerCode)

Return the length of the code.

CodingTheory.Xsigns — Method

Xsigns(S::CSSCode)

Return the signs of the X stabilizers of the CSS code.

CodingTheory.Xstabilizers — Method

Xstabilizers(S::CSSCode)

Return the X stabilizer matrix of the CSS code.

CodingTheory.Xsyndrome — Method

Xsyndrome(S::CSSCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the X stabilizers of the CSS code.

CodingTheory.Zsigns — Method

Zsigns(S::CSSCode)

Return the signs of the Z stabilizers of the CSS code.

CodingTheory.Zstabilizers — Method

Zstabilizers(S::CSSCode)

Return the Z stabilizer matrix of the CSS code.

CodingTheory.Zsyndrome — Method

Zsyndrome(S::CSSCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the Z stabilizers of the CSS code.

CodingTheory.allstabilizers — Function

allstabilizers(S::AbstractStabilizerCode, onlyprint::Bool=false)
elements(S::AbstractStabilizerCode, onlyprint::Bool=false)

Return the elements of S.

If onlyprint is true, the elements are only printed to the console and not returned.

CodingTheory.augment — Function

augment(S::AbstractStabilizerCode, row::fq_nmod_mat, symp::Bool=false, verbose::Bool=true)

Return the code created by added row to the stabilizers of S.

Notes:

The goal of this function is to track how the logical operators update given the new stabilizer. The unaffected logical operators are kept during the update and only those which don't commute with the new stabilizer are recomputed. Use verbose to better

CodingTheory.cardinality — Method

cardinality(S::AbstractStabilizerCode)

Return the cardinality of the stabilizer group of the code.

No size checking is done on the parameters of the code, returns a BitInt by default.

CodingTheory.changesigns! — Method

changesigns!(S::AbstractStabilizerCode, charvec::Vector{nmod})

Set the character vector of S to charvec and update the signs.

CodingTheory.charactervector — Method

charactervector(S::AbstractStabilizerCode)

Return the character vector of the code.

CodingTheory.dimension — Method

dimension(S::AbstractStabilizerCode)

Return the dimension of the code.

CodingTheory.expurgate — Function

expurgate(S::AbstractStabilizerCode, rows::Vector{Int}, verbose::Bool=true)

Return the code created by removing the stabilizers indexed by rows.

Notes:

The goal of this function is to track how the logical operators update through this process. Here, the original logical pairs are kept and an appropriate number of new pairs are added.

CodingTheory.field — Method

field(S::AbstractStabilizerCode)

Return the base ring of the code as a Nemo object.

CodingTheory.isCSS — Method

isCSS(S::AbstractStabilizerCode)

Return true is S is CSS.

Notes

This is intended to be a simple function wrapper for typeof(S) since the

constructor for QuantumCode automatically returns a CSSCode if possible. Manually changing the elements of the struct S without using the helper functions provided here is not recommended.

CodingTheory.logicalsmatrix — Method

logicalsmatrix(S::AbstractStabilizerCode)

Returns the result of logicals(S) as a vertically concatenated matrix.

CodingTheory.normalizermatrix — Method

normalizermatrix(S::AbstractStabilizerCode)

Return the normalizer matrix of the code.

CodingTheory.quadraticfield — Method

quadraticfield(S::AbstractStabilizerCode)

Return the quadratic field of the code as a Nemo object.

CodingTheory.rate — Method

rate(S::AbstractStabilizerCode)

Return the rate, `R = k/n', of the code.

CodingTheory.relativedistance — Method

relativedistance(S::AbstractStabilizerCode)

Return the relative minimum distance, δ = d / n of the code if d is known, otherwise errors.

CodingTheory.setlogicals! — Function

setlogicals!(S::AbstractStabilizerCode, L::fq_nmod_mat, symp::Bool=false)

Set the logical operators of S to L.

If the optional parameter symp is set to true, L is assumed to be in symplectic form over the base field of S.

CodingTheory.setminimumdistance! — Method

setminimumdistance(S::AbstractStabilizerCode, d::Int)

Set the minimum distance of the code to d.

The only check done on the value of d is that 1 ≤ d ≤ n.

CodingTheory.signs — Method

signs(S::AbstractStabilizerCode)

Return the signs of the stabilizers of the code.

CodingTheory.splitstabilizers — Method

splitstabilizers(S::AbstractStabilizerCode)

Return the set of X-only stabilizers and their signs, the set of Z-only stabilizers and their signs, and the remaining stabilizers and their signs.

This function returns six objevts of alternating types fq_nmod_mat and Vector{Int} for the three sets of stabilizers and signs, respectively. An empty set of stabilizers is returned as type Vector{fq_nmod_mat}.

CodingTheory.stabilizers — Method

stabilizers(S::AbstractStabilizerCode)

Return the stabilizer matrix of the code.

CodingTheory.symplecticstabilizers — Method

symplecticstabilizers(S::AbstractStabilizerCode)

Return the stabilizer matrix of the code in symplectic form.

CodingTheory.syndrome — Method

syndrome(S::AbstractStabilizerCode, v::fq_nmod_mat)

Return the syndrome of the vector v with respect to the stabilizers of S.