Tilings

CodingTheory.cosetintersection — Method

cosetintersection(genidxA::Vector{Int}, genidxB::Vector{Int}, subgroup::GapObj, g::ReflectionGroup)

Return the intersection of the cosets of g/subgroup wrt genidxA and wrt genidxB.

Notes

This outputs a sparse matrix with rows indexing the genidxA cosets and columns indexing the genidxB cosets.

CodingTheory.cycletetrahedrongroup — Method

cycletetrahedrongroup(q::Int, r::Int, s::Int, t::Int)

Return the "cycle" Coxeter group with high-order (>2) relations given by q, r, s, and t.

Corresponding Coxeter diagram:

   q
 o---o
t|   |r
 o---o
   s

CodingTheory.fixedpointfree — Method

fixedpointfree(subgroup::GapObj, g::ReflectionGroup)

Return true if the subgroup of g is fixed-point free; otherwise false.

CodingTheory.kcolorable — Method

kcolorable(k::Int, genidx::Vector{GapObj}, translations::Vector{GapObj}, subgroup::GapObj, g::ReflectionGroup)

Return true if the group elements corresponding to genidx in g/subgroup are k-colorable; otherwise false.

CodingTheory.normalsubgroups — Method

normalsubgroups(g::ReflectionGroup, maxindex::Int)

Return all normal subgroups of g with index up to maxindex.

CodingTheory.orientable — Method

orientable(subgroup::GapObj, g::ReflectionGroup)

Return true if the subgroup of g is orientable; otherwise false.

CodingTheory.qrsgroup — Method

qrsgroup(q::Int, r::Int, s::Int)

Return the Coxeter group corresponding to Schläfli symbol {q, r, s}.

Corresponding Coxeter diagram:

o---o---o---o
  q   r   s

CodingTheory.rsgroup — Method

rsgroup(r::Int, s::Int)

Return the Coxeter group corresponding to Schläfli symbol {r, s}.

Corresponding Coxeter diagram:

o---o---o
  r   s

CodingTheory.startetrahedrongroup — Method

startetrahedrongroup(q::Int, r::Int, s::Int)

Return the "star" Coxeter group with higher-order (>2) relations given by q, r, and s.

Corresponding Coxeter diagram:

      o
     / r
o---o
  q  \ s
      o

CodingTheory.tetrahedrongroup — Method

tetrahedrongroup(orders::Vector{Int})

Return the tetrahedron group with relations given by orders.

CodingTheory.trianglegroup — Method

trianglegroup(l::Int, m::Int, n::Int)

Return the (l, m, n) triangle group.