Tilings

CodingTheory.coset_intersection — Function

coset_intersection(gen_idx_A::Vector{Int}, gen_idx_B::Vector{Int}, subgroup::GapObj, g::ReflectionGroup)

Return the intersection of the cosets of g/subgroup wrt gen_idx_A and wrt gen_idx_B.

Notes

This outputs a sparse matrix with rows indexing the gen_idx_A cosets and columns indexing the gen_idx_B cosets.

CodingTheory.cycle_tetrahedron_group — Method

cycle_tetrahedron_group(q::Int, r::Int, s::Int, t::Int)

Return the "cycle" Coxeter group with high-order (>2) relations given by q, r, s, and t.

Corresponding Coxeter diagram:

   q
 o---o
t|   |r
 o---o
   s

CodingTheory.is_fixed_point_free — Method

is_fixed_point_free(subgroup::GapObj, g::ReflectionGroup)

Return true if the subgroup of g is fixed-point free; otherwise false.

CodingTheory.is_k_colorable — Method

is_k_colorable(k::Int, gen_idx::Vector{GapObj}, translations::Vector{GapObj}, subgroup::GapObj, g::ReflectionGroup)

Return true if the group elements corresponding to gen_idx in g/subgroup are k-colorable; otherwise false.

CodingTheory.is_orientable — Method

is_orientable(subgroup::GapObj, F::ReflectionGroup)

Return true if the subgroup of F is is_orientable; otherwise false.

CodingTheory.q_r_s_group — Method

q_r_s_group(q::Int, r::Int, s::Int)

Return the Coxeter group corresponding to Schläfli symbol {q, r, s}.

Corresponding Coxeter diagram:

o---o---o---o
  q   r   s

CodingTheory.r_s_group — Method

r_s_group(r::Int, s::Int)

Return the Coxeter group corresponding to Schläfli symbol {r, s}.

Corresponding Coxeter diagram:

o---o---o
  r   s

CodingTheory.star_tetrahedron_group — Method

star_tetrahedron_group(q::Int, r::Int, s::Int)

Return the "star" Coxeter group with higher-order (>2) relations given by q, r, and s.

Corresponding Coxeter diagram:

      o
     / r
o---o
  q  \ s
      o

CodingTheory.tetrahedron_group — Method

tetrahedron_group(orders::Vector{Int})

Return the tetrahedron group with relations given by orders.

CodingTheory.triangle_group — Method

triangle_group(l::Int, m::Int, n::Int)

Return the (l, m, n) triangle group.

Oscar.normal_subgroups — Method

normal_subgroups(g::ReflectionGroup, max_index::Int)

Return all normal subgroups of g with index up to max_index.