Trellises

CodingTheory.edges — Method

edges(T::Trellis)

Return the set of edges of the trellis T.

CodingTheory.vertices — Method

vertices(T::Trellis)

Return the set of vertices of the trellis T.