Trellises

CodingTheory.edges — Method

edges(T::Trellis)

Return the set of edges of the trellis T.

CodingTheory.isshifted — Method

isshifted(T::Trellis)

Return true if the trellis now represents a shifted version of the original code.

CodingTheory.trellisorientedformadditive — Method

trellisorientedformadditive(A::fq_nmod_mat)

Return the trellis oriented form of the matrix A assuming the row space is additive.

Note

So far this is only implemented for quadratic extensions over a prime subfield, i.e., GF(p^2).

CodingTheory.trellisorientedformlinear — Method

trellisorientedformlinear(A::fq_nmod_mat)

Return the trellis oriented form of the matrix A assuming the row space is linear.

CodingTheory.vertices — Method

vertices(T::Trellis)

Return the set of vertices of the trellis T.