Utilies

AbstractAlgebra.lift — Method

lift(A::MatElem{T}) where T <: ResElem

Return the matrix whose polynomial elements are converted to circulant matrices over the base field.

CodingTheory.:⊕ — Method

⊕(A::CTMatrixTypes, B::CTMatrixTypes)
directsum(A::CTMatrixTypes, B::CTMatrixTypes)

Return the direct sum of the two matrices A and B.

CodingTheory.:⊗ — Method

⊗(A::CTMatrixTypes, B::CTMatrixTypes)
kron(A::CTMatrixTypes, B::CTMatrixTypes)
tensorproduct(A::CTMatrixTypes, B::CTMatrixTypes)
kroneckerproduct(A::CTMatrixTypes, B::CTMatrixTypes)

Return the Kronecker product of the two matrices A and B.

CodingTheory.Hammingdistance — Method

Hammingdistance(u::T, v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer
distance(u::T, v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer
dist(u::T, v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer

Return the Hamming distance between u and v.

CodingTheory.Hammingweight — Method

Hammingweight(v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer
weight(v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer
wt(v::T) where T <: Union{CTMatrixTypes, Vector{S}} where S <: Integer

Return the Hamming weight of v.

CodingTheory.Hermitianconjugatematrix — Method

Hermitianconjugatematrix(A::CTMatrixTypes)

Return the Hermitian conjugate of the matrix A.

CodingTheory.Hermitianinnerproduct — Method

Hermitianinnerproduct(u::CTMatrixTypes, v::CTMatrixTypes)

Return the Hermitian inner product of u and v.

CodingTheory.areequivalentbasis — Method

areequivalentbasis(basis::Vector{fq_nmod}, basis2::Vector{fq_nmod})

Return true if basis is a scalar multiple of basis2.

CodingTheory.aresymplecticorthogonal — Method

aresymplecticorthogonal(A::CTMatrixTypes, B::CTMatrixTypes)

Return true if the rows of the matrices A and B are symplectic orthogonal.

CodingTheory.dualbasis — Method

dualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})
complementarybasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the dual (complentary) basis of basis for the extension E/F.

CodingTheory.edgevertexincidencegraph — Method

edgevertexincidencegraph(G::SimpleGraph{Int})

Return the edge-vertex incidence graph of G along with the vertex incides of the left and right bipartition.

CodingTheory.edgevertexincidencematrix — Method

edgevertexincidencematrix(G::SimpleGraph{Int})

Return the edge-vertex incidence matrix of G along with the vertex incides of the left and right bipartition.

CodingTheory.expandmatrix — Method

expandmatrix(M::CTMatrixTypes, K::FqNmodFiniteField, β::Vector{fq_nmod})

Return the matrix constructed by expanding the elements of M to the subfield K using the basis β for the base ring of M over K.

CodingTheory.extractbipartition — Method

extractbipartition(G::SimpleGraph{Int})

Return two vectors representing the vertex indices of each side of the bipartition.

CodingTheory.isbasis — Method

isbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true and the dual (complementary) basis if basis is a basis for E/F, otherwise return false, missing.

CodingTheory.isextension — Method

isextension(E::FqNmodFiniteField, F::FqNmodFiniteField)

Return true if E/F is a valid field extension.

CodingTheory.isnormalbasis — Method

isnormalbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is a normal basis for E/F.

CodingTheory.isprimitivebasis — Method

isprimitivebasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is a primitive basis for E/F.

CodingTheory.isselfdualbasis — Method

isselfdualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is equal to its dual.

CodingTheory.istriorthogonal — Function

istriorthogonal(G::CTMatrixTypes, verbose::Bool=false)
istriorthogonal(G::Matrix{Int}, verbose::Bool=false)

Return true if the binary matrix G is triorthogonal.

Notes

If the optional parameter verbos is set to true, the first pair or triple of non-orthogonal rows will be identified on the console.

CodingTheory.isvalidbipartition — Method

isvalidbipartition(G::SimpleGraph{Int}, left::Vector{Int}, right::Vector{Int})

Return true if the vertices indexed by left and right form a valid bipartition for G.

CodingTheory.polytocircmatrix — Method

polytocircmatrix(f::ResElem)

Return the circulant matrix whose first column is the coefficients of f.

CodingTheory.primitivebasis — Method

primitivebasis(E::FqNmodFiniteField, F::FqNmodFiniteField)

Return a primitive basis for E/F and its dual (complementary) basis.

CodingTheory.pseudoinverse — Method

pseudoinverse(M::CTMatrixTypes)

Return the pseudoinverse of a stabilizer matrix M over a quadratic extension.

Notes

This is not the Penrose-Moore pseudoinverse.

CodingTheory.quadraticresidues — Method

quadraticresidues(q::Int, n::Int)

Return the sets of quadratic resides and quadratic non-residues of q and n.

CodingTheory.rowsupports — Method

" rowsupports(M::CTMatrixTypes)

Returns a vector where the ith entry lists the indices of the nonzero entries of M[i, :]

CodingTheory.rowsupportssymplectic — Method

" rowsupportssymplectic(M::CTMatrixTypes)

Returns a vector where the ith entry is a 2-tuple of lists with the indices of the nonzero X and Z entries of M[i, :]

CodingTheory.symplecticinnerproduct — Method

symplecticinnerproduct(u::CTMatrixTypes, v::CTMatrixTypes)

Return the symplectic inner product of u and v.

CodingTheory.verifydualbasis — Method

verifydualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod}, dualbasis::Vector{fq_nmod})
verifycomplementarybasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod}, dualbasis::Vector{fq_nmod})

Return true if basis is the dual of dualbasis for E/F, otherwise return false.

CodingTheory.wt — Method

wt(f::CTPolyRingElem)

Return the number of nonzero coefficients of the polynomial f.

Hecke.isregular — Method

isregular(G::SimpleGraph{Int})

Return true if G is regular.

LinearAlgebra.tr — Function

tr(x::fq_nmod, K::FqNmodFiniteField, verify::Bool=false)

Return the relative trace of x from its base field to the field K.

Notes

If the optional parameter verify is set to true, the two fields are checked for compatibility.