Utilies

CodingTheory.:⊕ — Method

⊕(A::fq_nmod_mat, B::fq_nmod_mat)
directsum(A::fq_nmod_mat, B::fq_nmod_mat)

Return the direct sum of the two matrices A and B.

CodingTheory.:⊗ — Method

⊗(A::fq_nmod_mat, B::fq_nmod_mat)
kron(A::fq_nmod_mat, B::fq_nmod_mat)
tensorproduct(A::fq_nmod_mat, B::fq_nmod_mat)
kroneckerproduct(A::fq_nmod_mat, B::fq_nmod_mat)

Return the Kronecker product of the two matrices A and B.

CodingTheory.FpmattoJulia — Method

FpmattoJulia(M::fq_nmod_mat)

Return the fq_nmod_mat matrix M as a Julia Integer matrix.

CodingTheory.Hammingdistance — Method

Hammingdistance(u::T, v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer
distance(u::T, v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer
dist(u::T, v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer

Return the Hamming distance between u and v.

CodingTheory.Hammingweight — Method

Hammingweight(v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer
weight(v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer
wt(v::T) where T <: Union{fq_nmod_mat, Vector{S}} where S <: Integer

Return the Hamming weight of v.

CodingTheory.Hermitianconjugatematrix — Method

Hermitianconjugatematrix(A::fq_nmod_mat)

Return the Hermitian conjugate of the matrix A.

CodingTheory.Hermitianinnerproduct — Method

Hermitianinnerproduct(u::fq_nmod_mat, v::fq_nmod_mat)

Return the Hermitian inner product of u and v.

CodingTheory.aresymplecticorthogonal — Function

aresymplecticorthogonal(A::fq_nmod_mat, B::fq_nmod_mat, symp::Bool=false)

Return true if the rows of the matrices A and B are symplectic orthogonal.

If the optional parameter symp is set to true, A and B are assumed to be in symplectic form over the base field.

CodingTheory.dualbasis — Method

dualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})
complementarybasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the dual (complentary) basis of basis for the extension E/F.

CodingTheory.edgevertexincidencegraph — Method

edgevertexincidencegraph(G::SimpleGraph{Int})

Return the edge-vertex incidence graph of G along with the vertex incides of the left and right bipartition.

CodingTheory.edgevertexincidencematrix — Method

edgevertexincidencematrix(G::SimpleGraph{Int})

Return the edge-vertex incidence matrix of G along with the vertex incides of the left and right bipartition.

CodingTheory.expandmatrix — Method

expandmatrix(M::fq_nmod_mat, K::FqNmodFiniteField, basis::Vector{fq_nmod})

Return the matrix constructed by expanding the elements of M to the subfield K using the provided basis for the base ring of M over K.

CodingTheory.extractbipartition — Method

extractbipartition(G::SimpleGraph{Int})

Return two vectors representing the vertex indices of each side of the bipartition.

CodingTheory.isbasis — Method

isbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true and the dual (complementary) basis if basis is a basis for E/F, otherwise return false, missing.

CodingTheory.isequivalentbasis — Method

isequivalentbasis(basis::Vector{fq_nmod}, basis2::Vector{fq_nmod})

Return true if basis is a scalar multiple of basis2.`

CodingTheory.isextension — Method

isextension(E::FqNmodFiniteField, F::FqNmodFiniteField)

Return true if E/F is a valid field extension.

CodingTheory.isnormalbasis — Method

isnormalbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is a normal basis for E/F.

CodingTheory.isprimitivebasis — Method

isprimitivebasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is a primitive basis for E/F.

CodingTheory.isregular — Method

isregular(G::SimpleGraph{Int})

Return true if G is regular.

CodingTheory.isselfdualbasis — Method

isselfdualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod})

Return true if basis is equal to its dual.

CodingTheory.istriorthogonal — Function

istriorthogonal(G::fq_nmod_mat, verbose::Bool=false)
istriorthogonal(G::Matrix{Int}, verbose::Bool=false)

Return true if the binary matrix G is triorthogonal (modulo 2).

If the optional parameter verbos is set to true, the first pair or triple of non-orthogonal rows will be identified on the console.

CodingTheory.isvalidbipartition — Method

isvalidbipartition(G::SimpleGraph{Int}, left::Vector{Int}, right::Vector{Int})

Return true if the vertices indexed by left and right form a valid bipartition for G.

CodingTheory.polytocircmatrix — Method

polytocircmatrix(f::AbstractAlgebra.Generic.Res{fq_nmod_poly})

Return the circulant matrix whose first column is the coefficients of f.

CodingTheory.primitivebasis — Method

primitivebasis(E::FqNmodFiniteField, F::FqNmodFiniteField)

Return a primitive basis for E/F and its dual (complementary) basis.

CodingTheory.pseudoinverse — Method

pseudoinverse(M::fq_nmod_mat)

Return the pseudoinverse of a stabilizer matrix M over a quadratic extension.

Note that this is not the Penrose-Moore pseudoinverse.

CodingTheory.quadratictosymplectic — Method

quadratictosymplectic(M::fq_nmod_mat)

Return the matrix M converted from the quadratic to the symplectic form.

CodingTheory.symplecticinnerproduct — Method

symplecticinnerproduct(u::fq_nmod_mat, v::fq_nmod_mat)

Return the symplectic inner product of u and v.

CodingTheory.symplectictoquadratic — Method

symplectictoquadratic(M::fq_nmod_mat)

Return the matrix M converted from the symplectic to the quadratic form.

CodingTheory.verifydualbasis — Method

verifydualbasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod}, dualbasis::Vector{fq_nmod})
verifycomplementarybasis(E::FqNmodFiniteField, F::FqNmodFiniteField, basis::Vector{fq_nmod}, dualbasis::Vector{fq_nmod})

Return true if basis is the dual of dualbasis for E/F, otherwise return false.

CodingTheory.wt — Method

wt(f::fq_nmod_poly)

Return the number of nonzero coefficients of the polynomial f.

LinearAlgebra.tr — Function

tr(x::fq_nmod, K::FqNmodFiniteField, verify::Bool=false)

Return the relative trace of x from its base field to the field K.

If the optional parameter verify is set to true, the two fields are checked for compatibility.