Weight Enumerators, Distributions, And Minimum Distances

AbstractAlgebra.polynomial — Method

polynomial(W::WeightEnumerator)

Returns the polynomial of the weight enumerator W.

CodingTheory.Sternsattack — Function

Sternsattack(C::AbstractLinearCode, w::Int, p::Int, l::Int)

Search for codewords of C of weight w using Stern's attach and return any found.

CodingTheory.minimumdistance — Function

minimumdistance(Q::AbstractStabilizerCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the stabilizer code if known, otherwise computes it.

CodingTheory.minimumdistance — Function

minimumdistance(C::AbstractLinearCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the linear code if known, otherwise computes it using the algorithm of alg. If alg = "trellis", the sectionalization flag sect can be set to true to further compactify the reprsentation.

enumeration-based algorithms using the binary, reflected Gray code such as Brouwer-Zimmermann and adaptions

CodingTheory.support — Function

support(S::AbstractStabilizerCode, alg::String="auto", type::String="stabilizer")

Returns the support related to S.

The support is the collection of nonzero exponents of the Hamming weight enumerator. If type is stabilizer, the support of the stabilizers are computed. If type is normalizer, the support of the normalizer of the stabilizers are computed. If type is quotient, the support of the normalizer mod the stabilizers (logical representatives only) is computed.

CodingTheory.support — Method

support(C::AbstractLinearCode)

Returns the support of C.

The support of C is the collection of nonzero exponents of the Hamming weight enumerator of C.

CodingTheory.type — Method

type(W::WeightEnumerator)

Returns the type of the weight enumerator W.

CodingTheory.weightplot — Function

weightplot(S::AbstractStabilizerCode, alg::String="auto", type::String="stabilizer")

Return a bar plot of the weight distribution related to S.

If type is stabilizer, the weight distribution of the stabilizers are computed. If type is normalizer, the weight distrbution of the normalizer of the stabilizers are computed. If type is quotient, the weight distrbution of the normalizer mod the stabilizers (logical representatives only) is computed.

CodingTheory.weightplot — Function

weighthplot(C::AbstractLinearCode, alg::String="auto")

Return a bar plot of the weight distribution of C.

CodingTheory.weightplotCSS — Function

weightplotCSS(S::AbstractCSSCode, alg::String="auto")

Return bar plots of the weight distribution of the both the X and 'Z' stabilizers, separately.

CodingTheory.weightplotCSSX — Function

weightplotCSSX(S::AbstractCSSCode, alg::String="auto")

Return a bar plot of the weight distribution of the X stabilizers.

CodingTheory.weightplotCSSZ — Function

weightplotCSSZ(S::AbstractCSSCode, alg::String="auto")

Return a bar plot of the weight distribution of the Z stabilizers.