Weight Enumerators, Distributions, And Minimum Distances

AbstractAlgebra.polynomial — Method

polynomial(W::WeightEnumerator)

Returns the polynomial of the weight enumerator W.

CodingTheory.minimumdistance — Function

minimumdistance(C::AbstractLinearCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the linear code if known, otherwise computes it using the algorithm of alg. If alg = "trellis", the sectionalization flag sect can be set to true to further compactify the reprsentation.

CodingTheory.minimumdistance — Method

minimumdistance(Q::AbstractStabilizerCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the stabilizer code if known, otherwise computes it.

CodingTheory.support — Method

support(C::AbstractLinearCode)

Returns the support of C.

The support of C is the collection of nonzero exponents of the Hamming weight enumerator of C.

CodingTheory.type — Method

type(W::WeightEnumerator)

Returns the type of the weight enumerator W.