Weight Enumerators, Distributions, And Minimum Distances

For more information on the algorithms contained in this function see "Enumeration Based Algorithms" by Gregory White.

enumeration-based algorithms using the binary, reflected Gray code such as Brouwer-Zimmermann and adaptions

notational used for complete and Hamming weight enumerators

AbstractAlgebra.polynomial — Method

polynomial(W::WeightEnumerator)

Returns the polynomial of the weight enumerator W.

CodingTheory.CWEtoHWE — Method

CWEtoHWE(CWE::WeightEnumerator)

Return the Hamming weight enumerator associated with the complete weight enumerator CWE.

CodingTheory.Sternsattack — Function

Sternsattack(C::AbstractLinearCode, w::Int, p::Int, l::Int)

Search for codewords of C of weight w using Stern's attack and return any found.

CodingTheory.minimumdistance — Function

minimumdistance(C::AbstractLinearCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the linear code if known, otherwise computes it using the algorithm of alg. If alg = "trellis", the sectionalization flag sect can be set to true to further compactify the reprsentation.

CodingTheory.minimumdistance — Function

minimumdistance(Q::AbstractStabilizerCode, alg::String="trellis", sect::Bool=false)

Return the minimum distance of the stabilizer code if known, otherwise computes it.

CodingTheory.minimumwords — Method

minimumwords(C::AbstractLinearCode)

Return the set of codewords of C with weight equal to the minimum distance.

Notes

This algorithm simultaneously computes the minimum distance and stores the words of this weight that it finds, removing the repeated work of calling w = minimumdistance(C); W = wordsofweight(C, w);

CodingTheory.type — Method

type(W::WeightEnumerator)

Returns the type of the weight enumerator W.

CodingTheory.weightdistribution — Function

weightdistribution(C::AbstractLinearCode, alg::String="auto", compact::Bool=true)

Return the weight distribution of C using the algorithm alg. If compact is false, the result will be a Vector{BigInt} of length length(C) + 1 whose ith entry is the number of codewords of C of Hamming weight i - 1. Otherwise, the result is a Vector{Tuple{Int, BigInt}} whose entries specify the nonzero indices and values of the above.

CodingTheory.weightenumerator — Function

weightenumerator(C::AbstractLinearCode, type::Symbol=:complete, alg::String="auto")

Return either the :complete or :Hamming weight enumerator of C using the algorithm alg.

CodingTheory.weightplot — Function

weightplot(C::AbstractLinearCode, alg::String="auto")

Return a bar plot of the weight distribution of C.

CodingTheory.weightplot — Function

weightplot(S::AbstractStabilizerCode, alg::String="auto", type::String="stabilizer")

Return a bar plot of the weight distribution related to S.

If type is stabilizer, the weight distribution of the stabilizers are computed. If type is normalizer, the weight distrbution of the normalizer of the stabilizers are computed. If type is quotient, the weight distrbution of the normalizer mod the stabilizers (logical representatives only) is computed.

CodingTheory.weightplotCSS — Function

weightplotCSS(S::AbstractStabilizerCodeCSS, alg::String="auto")

Return bar plots of the weight distribution of the both the X and 'Z' stabilizers, separately.

CodingTheory.weightplotCSSX — Function

weightplotCSSX(S::AbstractStabilizerCodeCSS, alg::String="auto")

Return a bar plot of the weight distribution of the X stabilizers.

CodingTheory.weightplotCSSZ — Function

weightplotCSSZ(S::AbstractStabilizerCodeCSS, alg::String="auto")

Return a bar plot of the weight distribution of the Z stabilizers.

CodingTheory.wordsofweight — Function

wordsofweight(C::AbstractLinearCode, lbound::Int, ubound::Int, verbose::Bool=false)

Return all the codewords of C of Hamming weight in the range [lbound, ubound].

CodingTheory.wordsofweight — Function

wordsofweight(C::AbstractLinearCode, bound::Int, verbose::Bool=false)

Return all the codewords of C of Hamming weight bound.

Hecke.support — Function

support(S::AbstractStabilizerCode, alg::String="auto", type::String="stabilizer")

Returns the support related to S.

The support is the collection of nonzero exponents of the Hamming weight enumerator. If type is stabilizer, the support of the stabilizers are computed. If type is normalizer, the support of the normalizer of the stabilizers are computed. If type is quotient, the support of the normalizer mod the stabilizers (logical representatives only) is computed.

Hecke.support — Method

support(C::AbstractLinearCode)

Returns the support of C.

Notes

The support of C is the collection of nonzero exponents of the Hamming weight enumerator of C.